% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.global.climate.R
\name{build.global.climate}
\alias{build.global.climate}
\title{build.global.climate}
\usage{
get.global.climate(folder)
}
\arguments{
\item{folder}{Path to the folder you want to install the global climate data in}
}
\description{
Downloads and builds the global climate, elevation and soil moisture datasets used in the function micro_global. The global
climate dataset global.nc is generated from "New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution
data set of surface climate over global land areas. Climate Research 21:1-25", specically 10' estimates of monthly
mean (over 1961-1990) preceipitation, wet-days, air temperature, diurnal temperature range, relative humidity and
wind speed. Cloud cover comes from a bilinear interpolation of a lower resolution version of this dataset (New, M.,
M. Hulme, and P. D. Jones. 1999. Representing twentieth century space-time climate variability. Part 1: development
of a 1961-90 mean monthly terrestrial climatology. Journal of Climate 12:829-856.). The micro_global function
optionaly uses a  global monthly soil moisture estimate from NOAA CPC Soil Moisture http://140.172.38.100/psd/
thredds/catalog/Datasets/cpcsoil/catalog.html
}
