% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CampNormTbl9_1.R
\docType{data}
\name{CampNormTbl9_1}
\alias{CampNormTbl9_1}
\title{Table 9.1 from Campbell and Norman 1990.}
\format{
A data frame with 11 rows and 8 variables:
\describe{
  \item{Texture}{soil type}
  \item{Silt}{proportion silt}
  \item{Clay}{proportion clay}
  \item{airentry}{air entry potential, in J/kg}
  \item{b}{b parameter, in -}
  \item{Ks}{saturated conductivity, in kg s m-3}
  \item{field}{field capacity (-1500 J/kg) water content, in m3/m3}
  \item{wilting}{wilting point (-33 J/kg) water content, in m3/m3}
}
}
\source{
Campbell, G. S., and J. M. Norman. 1998. Environmental Biophysics. Springer, New York.
}
\usage{
CampNormTbl9_1
}
\description{
Hydraulic properties of soils as a function of soil texture
}
\keyword{datasets}
