% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONV_ecto.R
\encoding{UTF-8}
\name{CONV_ecto}
\alias{CONV_ecto}
\title{CONV_ecto}
\usage{
CONV_ecto(
  GEOMETRY = 3,
  ATOT = 0.01325006,
  AV = 0.001325006,
  AL = 0.03419952,
  AT = 0,
  BP = 101325,
  ALT = 0,
  TA = 20,
  VEL = 1,
  FLTYPE = 0,
  TSKIN = 25
)
}
\arguments{
\item{GEOMETRY}{organism shape, 0-5, 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom}

\item{ATOT}{total body surface area (m2)}

\item{AV}{ventral surface area (m2)}

\item{AL}{animal characteristic dimension (length) (m)}

\item{AT}{body surface area contacting another organism of same temperature (m2)}

\item{BP}{air pressure (Pa)}

\item{ALT}{elevation (m)}

\item{TA}{air temperature (°C)}

\item{VEL}{wind speed (m/s)}

\item{FLTYPE}{fluid type, air (0) or water (1)}

\item{TSKIN}{skin temperature (°C)}
}
\description{
R version of Fortran CONV.f (ectotherm model) for calculating convective exchange and heat and mass transfer coefficients.
}
