% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONV.R
\name{CONV_ENDO}
\alias{CONV_ENDO}
\title{CONV_ENDO}
\usage{
CONV_ENDO(TS, TENV, NGEOM, SURFAR, FLTYPE, FURTST, D, TFA, VEL, ZFUR, BP,
  ELEV)
}
\arguments{
\item{TS}{A}

\item{TENV}{A}

\item{NGEOM}{A}

\item{SURFAR}{A}

\item{FLTYPE}{A}

\item{FURTST}{A}

\item{D}{A}

\item{TFA}{A}

\item{VEL}{A}

\item{ZFUR}{A}

\item{BP}{A}

\item{ELEV}{A}
}
\description{
R wrapper for Fortran binary of CONV_ENDO (endotherm model)
}
