% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cosby1984Tbl4.R
\docType{data}
\name{Cosby1984Tbl4}
\alias{Cosby1984Tbl4}
\title{Table 4 from Cosby et al. (1984).}
\format{A data frame with 8 rows and 5 variables:
\describe{
  \item{parameter}{hydraulic parameter}
  \item{intercept}{intercept of regression}
  \item{variable}{soil component to act as dependent variable}
  \item{slope}{slope of regression}
  \item{R2}{R2 of regression}
  \item{deltaR2}{delta R2 of regression}
  \item{p}{p value of regression}
  \item{n}{sample size regression}
}}
\source{
Cosby, B. J., G. M. Hornberger, R. B. Clapp, and T. R. Ginn. 1984. A Statistical Exploration of the Relationships of Soil Moisture Characteristics to the Physical Properties of Soils. Water Resources Research 20:682-690.
}
\usage{
Cosby1984Tbl4
}
\description{
Paramters to obtain hydraulic properties of soils as a function of soil composition
}
\keyword{datasets}
