% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crudat2raster.R
\name{crudat2raster}
\alias{crudat2raster}
\title{crudat2raster}
\usage{
get.global.climate(folder)
}
\arguments{
\item{folder}{Path to the folder you want to install the global climate data in}
}
\value{
cru.raster CRU data in raster format

header Header data for the layer
}
\description{
Converts CRU integer ascii file to raster with longitude range -180 to 180
}
