% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectoR_devel.R
\encoding{UTF-8}
\name{ectoR_devel}
\alias{ectoR_devel}
\title{Ectotherm heat budget model}
\usage{
ectotherm(Ww_g = 40, shape = 3, alpha = 0.85, postur = 0, TA = 20, TGRD = 40, TSKY = -5, VEL = 1, RH = 30, QSOLR = 800, Z = 20 ...)
}
\arguments{
\item{Ww_g}{= 40, Wet weight of animal (g), note this model is 'steady state' so no lags in heating/cooling due to mass}

\item{alpha}{= 0.85, Solar absorptivity, 0-1}

\item{shape}{= 3, Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see details)}

\item{M_1}{= 0.013, Metabolic rate parameter 1 V_O2=M_1*M^M_2*10^(M_3*Tb), in ml O2 / h, default parameters for lizards based on Eq. 2 from Andrews & Pough 1985. Physiol. Zool. 58:214-231}

\item{M_2}{= 0.800, Metabolic rate parameter 2}

\item{M_3}{= 0.038, Metabolic rate parameter 3}

\item{pct_wet}{= 0.1, \% of surface area acting as a free-water exchanger, for computing cutaneous water loss}

\item{pct_eyes}{= 0, \% of surface area taken up by open eyes, for computing ocular water loss (only when active)}

\item{pct_mouth}{= 0, \% of surface area taken up by open mouth, for computing panting water loss}

\item{psi_body}{= -7.07 * 100, water potential of body (J/kg) - affects skin humidity for water vapour exchange}

\item{pantmax}{= 1, maximum multiplier on breathing rate, for respiratory water loss via panting (>1 invokes panting)}

\item{F_O2}{= 20, \% oxygen extraction efficiency, for respiratory water loss}

\item{delta_air}{= 0.1, temperature difference (°C) between expired and inspired air, for computing respiratory water loss}

\item{O2gas}{= 20.95, \% O2 in air}

\item{CO2gas}{= 0.03, \% CO2 in air}

\item{N2gas}{= 79.02, \% nitrogen in air}

\item{SHADE}{= 0, shade level (\%)}

\item{QSOLR}{= 800, solar radiation, horizontal plane (W/m2)}

\item{Z}{= 20, zenith angle of sun (degrees from overhead)}

\item{TA}{= 20, air temperature at local height (°C)}

\item{TGRD}{= 40, ground temperature (°C)}

\item{TSKY}{= -5, sky temperature (°C)}

\item{VEL}{= 1, wind speed (m/s)}

\item{RH}{= 30, relative humidity (\%)}
}
\description{
A modularised implementation of the heat budget of the Niche Mapper ectotherm model that computes body temperature,
and water loss for a single environment with no behaviour. Like endoR_devel, it can be used to develop customised
versions of the NicheMapR ectotherm model.
}
\details{
\strong{ Environmental inputs:}

\itemize{
\item{\code{elev}{ = 0, elevation (m)}\cr}
\item{\code{alpha_sub}{ = 0.8, solar absorptivity of substrate (fractional, 0-1)}\cr}
\item{\code{fluid}{ = 0, fluid type: 0 = air; 1 = water}\cr}
\item{\code{TSUBST}{ = TGRD, surface temperature for conduction (°C)}\cr}
\item{\code{K_sub}{ = 2.79, substrate thermal conductivity (W/m°C)}\cr}
\item{\code{pres}{ = 101325, atmospheric pressure (Pa)}\cr}
\item{\code{O2gas}{ = 20.95, oxygen concentration of air, to account for non-atmospheric concentrations e.g. in burrows (\%)}\cr}
\item{\code{N2gas}{ = 79.02, nitrogen concetration of air, to account for non-atmospheric concentrations e.g. in burrows (\%)}\cr}
\item{\code{CO2gas}{ = 0.0412, carbon dioxide concentration of air, to account for non-atmospheric concentrations e.g. in burrows (\%)}\cr}
\item{\code{PDIF}{ = 0.15, proportion of solar radiation that is diffuse (fractional, 0-1)}\cr}
}
\strong{ Morphological parameters:}

\itemize{
\item{\code{custom_shape}{ = c(10.4713,.688,0.425,0.85,3.798,.683,0.694,.743), Custom shape coefficients. Operates if shape=5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*mass^b, where AREA is in cm2 and mass is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}\cr}
\item{\code{shape_a}{ = 1, Proportionality factor (-) for going from volume to area, keep this 1 (redundant parameter that should be removed)}\cr}
\item{\code{shape_b}{ = 3, Proportionality factor (-) for going from volume to area, represents ratio of width:height for a plate, length:diameter for cylinder, b axis:a axis for ellipsoid }\cr}
\item{\code{shape_c}{ = 0.6666666667, Proportionality factor (-) for going from volume to area, represents ratio of length:height for a plate, c axis:a axis for ellipsoid}\cr}
\item{\code{fatosk}{ = 0.4, Configuration factor to sky (-) for infrared calculations}\cr}
\item{\code{fatosb}{ = 0.4, Configuration factor to subsrate for infrared calculations}\cr}
\item{\code{pct_cond}{ = 10, Percentage of animal surface contacting the substrate (\%)}\cr}
\item{\code{pct_touch}{ = 0, Percentage of animal surface area contacting another animal of same temperature (\%)}\cr}
\item{\code{k_flesh}{ = 0.5, Thermal conductivity of flesh (W/mC, range: 0.412-2.8)}\cr}
\item{\code{rho_body}{ = 1000, Density of flesh (kg/m3)}\cr}
\item{\code{epsilon}{ = 0.95, Emissivity of animal (0-1)}\cr}
\item{\code{postur}{ = 1, postural orientation to sun, 1 = perpendicular, 2 = parallel, 0 = half way between (foraging)}\cr}
}
\strong{Outputs:}

temperature variables:

TC - Body temperature (°C)

TSKIN - Skin temperature (°C)

TLUNG - Lung temperature (°C)

enbal variables:
\itemize{
\item 1 QSOL - Solar radiation absorbed (W)
\item 2 QIRIN - Infrared radiation absorbed (W)
\item 3 QMET - Metabolic heat production (W)
\item 4 QEVAP - Evaporative heat loss (W)
\item 5 QIROUT - Infrared radiation lost (W)
\item 6 QCONV - Heat lost by convection (W)
\item 7 QCOND - Heat lost by conduction (W)
\item 8 ENB - Energy balance (W)
\item 9 NTRY - Iterations that were required for solution to heat balance equation
}
masbal variables:
\itemize{
\item 1 O2_ml - Oxygen consumption rate (ml/h)
\item 2 H2OResp_g - Respiratory water loss (g/h)
\item 3 H2OCut_g - Cutaneous water loss (g/h)
\item 4 H2OEye_g - Ocular water loss (g/h)
}
}
\examples{
library(NicheMapR)

# parameters
Ww_g <- 40
shape <- 3
alpha <- 0.85
ectoR.out <- ectoR_devel(Ww_g = Ww_g, # wet weight, g
                         shape = shape, # using lizard geometry
                         alpha = alpha, # solar absorptivity
                         postur = 0, # average posture, half way between normal and parallel to sun
                         TA = 20, # air temperature at lizard height, deg C
                         TGRD = 40, # ground temperature, deg C
                         TSKY = -5, # sky temperature, deg C
                         VEL = 1, # wind speed, m/s
                         RH = 30, # relative humidity, \%
                         QSOLR = 800, # total horizontal plane solar radiation, W/m2
                         Z = 20 # solar zenith angle, degrees
)
# return body temperature
ectoR.out$TC
# return skin temperature
ectoR.out$TS
# return heat budget
ectoR.out$enbal
# return O2 consumption rate and water loss
ectoR.out$masbal

# run microclimate model in monthly mode at default site (Madison, Wisconsin, USA)
micro <- micro_global()

# extract full sun conditions
metout <- as.data.frame(micro$metout)
soil <- as.data.frame(micro$soil)

# get required inputs
TAs <- metout$TALOC
TGRDs <- soil$D0cm
TSKYs <- metout$TSKYC
VELs <- metout$VLOC
RHs <- metout$RHLOC
QSOLRs <- metout$SOLR
Zs <- metout$ZEN

# use ectoR_devel to compute body temperature in open without respiratory heat loss,
# conduction, or metabolic heat gain
TC <- unlist(lapply(1:length(TAs), function(x){ectoR_devel(
  Ww_g = Ww_g, # wet weight, g
  shape = shape, # using lizard geometry
  alpha = alpha, # solar absorptivity
  M_1 = 0, # turn of metabolic heat
  postur = 0, # average posture, half way between normal and parallel to sun
  pantmax = 0, # turn off respiratory heat exchange
  pct_cond = 0, # negligible conduction to substrate
  alpha_sub = (1 - micro$REF), # substrate absorptivity used in microclimate model
  elev = micro$elev, # elevation from microclimate model
  TA = TAs[x], # air temperature at lizard height from microclimate model, deg C
  TGRD = TGRDs[x], # ground temperature from microclimate model, deg C
  TSKY = TSKYs[x], # sky temperature from microclimate model, deg C
  VEL = VELs[x], # wind speed from microclimate model, m/s
  RH = RHs[x], # relative humidity from microclimate model, \%
  QSOLR = QSOLRs[x], # total horizontal plane solar radiation from microclimate model, W/m2
  Z = Zs[x] # solar zenith angle from microclimate model, degrees
)$TC})) # run ectoR_devel across environments

# run ectotherm model for a non-behaving animal without respiratory heat loss,
# conduction or metabolic heat gain
ecto <- ectotherm(
  Ww_g = Ww_g, # wet weight, g
  shape = shape, # using lizard geometry
  alpha_min = alpha, # minimum solar absorptivity
  alpha_max = alpha, # maximum solar absorptivity
  M_1 = 0, # turn of metabolic heat
  postur = 0, # average posture, half way between normal and parallel to sun
  pantmax = 0, # turn off respiratory heat exchange
  pct_cond = 0, # negligible conduction to substrate
  live = 0
)

# extract results
environ <- as.data.frame(ecto$environ)
enbal <- as.data.frame(ecto$enbal)
masbal <- as.data.frame(ecto$masbal)
TC_ectotherm <- environ$TC

# compare
time <- micro$dates
plot(time, TC, type = 'l', ylim = c(-25, 55), ylab = 'body temperature, deg C', xlab = 'month of year')
points(time, TC_ectotherm, type = 'l', col = 2)

}
