% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectorun.R
\name{ectorun}
\alias{ectorun}
\title{ectorun}
\usage{
ectorun(ecto)
}
\arguments{
\item{ecto}{A vector of input variables for the microclimate model}

\item{metout}{The above ground micrometeorological conditions under the minimum specified shade}

\item{shadmet}{The above ground micrometeorological conditions under the maximum specified shade}

\item{soil}{Hourly predictions of the soil temperatures under the minimum specified shade}

\item{shadsoil}{Hourly predictions of the soil temperatures under the maximum specified shade}

\item{soilmoist}{Hourly predictions of the soil moisture under the minimum specified shade}

\item{shadmoist}{Hourly predictions of the soil moisture under the maximum specified shade}

\item{soilpot}{Hourly predictions of the soil water potential under the minimum specified shade}

\item{shadpot}{Hourly predictions of the soil water potential under the maximum specified shade}

\item{humid}{Hourly predictions of the soil humidity under the minimum specified shade}

\item{shadhumid}{Hourly predictions of the soil humidity under the maximum specified shade}

\item{DEP}{Depths used for the microclimate model}

\item{rainfall}{Daily rainfall}

\item{rainhr}{Hourly rainfall (overwrites rainfall if non-negative}

\item{debmod}{Dynamic Energy Budget (DEB) model paramters}

\item{deblast}{Initial DEB state}

\item{foodwaters}{Food water content (add units)}

\item{foodlevels}{Food levels (add units)}

\item{wetlandTemps}{Tempature of water body}

\item{wetlandDepths}{Depth of water body}

\item{GLMtemps}{General Lake Model water temperatures}

\item{GLMO2s}{General Lake Model water O2 partial pressures}

\item{GLMsalts}{General Lake Model water salinities}

\item{GLMpHs}{General Lake Model water pHs}

\item{GLMfoods}{General Lake Model water food densities}

\item{arrhenius}{Arrhenius thermal response}

\item{thermal_stages}{Stage-specific thermal physiology}

\item{behav_stages}{Stage-specific behaviour}

\item{water_stages}{Stage-specific water related parameters}

\item{nutri_stages}{Stage-specific nutrition related parameters}

\item{MINSHADES}{Minimum shade levels}

\item{MAXSHADES}{Maximum shade levels}

\item{S_instar}{For the Insect DEB model}
}
\description{
R wrapper for Fortran binary of Niche Mapper microclimate model
}
