% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectotherm.R
\name{ectotherm}
\alias{ectotherm}
\title{ectotherm}
\usage{
ectotherm(amass, lometry, ABSMAX, ABSMIN, TMAXPR, TMINPR, TBASK, TEMERGE, ctmax, ctmin,
 tpref, dayact, nocturn, crepus, CkGrShad, burrow, climb, shdburrow, mindepth, maxdepth,
 MR_1, MR_2, MR_3, ...)
}
\arguments{
\item{amass}{Mass of animal (g), note this model is 'steady state' so no lags in heating/cooling due to mass}

\item{lometry}{Organism shape, 0-4 (see details)}

\item{ABSMAX}{Maximum solar absorptivity, decimal percent}

\item{ABSMIN}{Maximum solar absorptivity, decimal percent}

\item{TMAXPR}{Voluntary thermal maximum, degrees C (upper body temperature for foraging and also affects burrow depth selection)}

\item{TMINPR}{Voluntary thermal minimum, degrees C (lower body temperature for foraging)}

\item{TBASK}{Minimum basking temperature, degrees C}

\item{TEMERGE}{Temperature at which animal will move to a basking site, degrees C}

\item{ctmax}{Critical thermal maximum, degrees C (affects burrow selection)}

\item{ctmin}{Critical thermal minimum, degrees C (affects burrow selection)}

\item{dayact}{Diurnal activity allowed (1) or not (0)?}

\item{nocturn}{Nocturnal activity allowed (1) or not (0)?}

\item{crepus}{Crepuscular activity allowed (1) or not (0)?}

\item{CkGrShad}{shade seeking allowed (1) or not (0)?}

\item{burrow}{Shelter in burrow allowed (1) or not (0)?}

\item{climb}{climbing to seek cooler habitats allowed (1) or not (0)?}

\item{shdburrow}{choose if the animal's retreat is in the shade (1) or in the open (0)}

\item{mindepth}{Minimum depth (soil node #) to which animal can retreat if burrowing}

\item{maxdepth}{Maximum depth (soil node #) to which animal can retreat if burrowing}

\item{MR_1}{Metabolic rate parameter MR=MR_1*M^MR_2*10^(MR_3*Tb) based on Eq. 2 from Andrews & Pough 1985. Physiol. Zool. 58:214-231}

\item{MR_2}{Metabolic rate parameter}

\item{MR_3}{Metabolic rate parameter}

\item{tpref}{Preferred body temperature, degrees C}
}
\description{
An implementation of the Niche Mapper ectotherm model
}

