% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_endo.R
\name{ellipsoid}
\alias{ellipsoid}
\title{Implementation of the endotherm model .}
\usage{
ellipsoid(posture = 4.5, mass = 0.5, coreT = 37, furdepth = 5,
  O2eff = 0.2, furcond = 0.04, airT = 20, windspd = 1, rh = 50)
}
\arguments{
\item{posture}{Shape, ratio of long to short axis of a prolate ellipsoid}

\item{mass}{Body Mass (kg)}

\item{coreT}{Core temperature (deg C)}

\item{furdepth}{Fur depth (mm)}

\item{O2eff}{Oxygen extraction efficiency (decimal \%)}

\item{furcond}{Conductivity of fur (W/Cm)}

\item{airT}{Air temperature (deg C)}

\item{windspd}{Wind speed (m/s)}

\item{rh}{Relative humidity (\%)}
}
\value{
AirTemp

Windspeed

RelHum

Tcore

UpperCritTemp

LowerCritTemp

Qresp_gph, respiration rate in grams per hour

Qresp_W, respiration rate in watts

Qresp_kjph, respiratino rate in kilojules per hour

Tskin, skin temp

Qgen, required heat generation

QgenFinal, required heat generation capped at basal

mlO2ph, mls of O2 consumed per hour

PctBasal, metabolic rate expressed as a percentage of basal

status, 1=cold, 2=happy, 3=hot

H2Oloss_W, watts of heat to be dumped

H2O_gph, water loss rate required to dump heat based on latent heat of vaporization

massph_percent, percent of body mass lost as water per hour

timetodeath, time to death (hours) from desiccation (15\% desiccated) if no water to drink
}
\description{
An implementation of described in
Porter, W. P., and M. Kearney. 2009. Size, shape, and the thermal niche of endotherms.
Proceedings of the National Academy of Sciences 106:19666-19672.
with additional rough calculations for water loss. Note this model is only relevant for
conditions where there is no solar radiation and air/ground/sky temperatures are equal
}

