% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endoR.R
\encoding{UTF-8}
\name{endoR}
\alias{endoR}
\title{endoR}
\usage{
endoR(AMASS = 1, NGEOM = 4, GMREF = 3, FURTHRMK = 0, ZFURD = 2E-03, ZFURV = 2E-03, TC = 37, TCMAX = 45, TA = 20, TGRD = TA, TSKY = TA, VEL = 0.1, RH = 5, QSOLR = 0, Z = 20, SHADE = 0, NITESHAD = 0,...)
}
\arguments{
\item{TA}{= 20, air temperature at local height (°C)}

\item{TGRD}{= TA, ground temperature (°C)}

\item{TSKY}{= TA, sky temperature (°C)}

\item{VEL}{= 0.1, wind speed (m/s)}

\item{RH}{= 5, relative humidity (\%)}

\item{QSOLR}{= 0, solar radiation, horizontal plane (W/m2)}

\item{Z}{= 20, zenith angle of sun (degrees from overhead)}

\item{SHADE}{= 0, shade level (\%)}

\item{AMASS}{= 1, # kg}

\item{NGEOM}{= 4, # cylinder (ngeom = 1), sphere (ngeom = 2) and ellipsoid (ngeom = 4). If a truncated cone (5) or ellipsoidal cylinder (3), we will use the cylinder equations (ngeom=1).}

\item{GMREF}{= 3, # initial ratio between long and short axis (-)}

\item{FURTHRMK}{= 0, # user-specified fur thermal conductivity, not used if 0}

\item{ZFURD}{= 2E-03, # fur depth, dorsal (m)}

\item{ZFURV}{= 2E-03, # fur depth, ventral (m)}

\item{TC}{= 37, # core temperature (°C)}

\item{TCMAX}{= 45, # maximum core temperature (°C)}
}
\description{
Endotherm model of NicheMapR
}
\details{
\strong{ Parameters controlling how the model runs:}\cr\cr
\code{DIFTOL}{ = 0.001, error tolerance for SIMULSOL (W)}\cr\cr

\strong{ Environment:}\cr\cr
\code{TAREF}{ = TA, air temperature at reference height (°C)}\cr\cr
\code{ELEV}{ = 0, elevation (m)}\cr\cr
\code{ABSSB}{ = 0.8, solar absorptivity of substrate (dec \%)}\cr\cr
\code{FLTYPE}{ = 0, FLUID TYPE: 0 = AIR; 1 = FRESH WATER; 2 = SALT WATER - need's to be looked at - only invoked in main program when the dive table is set up}\cr\cr
\code{TCONDSB}{ = TGRD, surface temperature for conduction (°C)}\cr\cr
\code{TBUSH}{ = TA, bush temperature (°C)}\cr\cr
\code{BP}{ = -1, Pa, negatve means elevation is used}\cr\cr
\code{O2GAS}{ = 20.95, oxygen concentration of air (\%)}\cr\cr
\code{N2GAS}{ = 79.02, nitrogen concetration of air (\%)}\cr\cr
\code{CO2GAS}{ = 0.03, carbon dioxide concentration of air (\%)}\cr\cr
\code{PCTDIF}{ = 0.15, proportion of solar radiation that is diffuse (dec \%)}\cr\cr

\strong{ Behaviour:}\cr\cr
\code{SHADE}{ = 0, shade level (\%)}\cr\cr
\code{NITESHAD}{ = 0, flag for if animal is behaviourally seeking shade for warmth at night - remove?}\cr\cr
\code{FLYHR}{ = 0, is flight occuring this hour? (imposes forced evaporative loss)}\cr\cr
\code{UNCURL}{ = 0.1, allows the animal to uncurl to GMULTMAX, the value being the increment GMULT is increased per iteration}\cr\cr
\code{RAISETC}{ = 0.1, turns on core temperature elevation, the value being the increment by which TC is increased per iteration}\cr\cr
\code{SWEAT}{ = 0.1, turns on sweating, the value being the increment by which SKINW is increased per iteration (\%)}\cr\cr
\code{MXWET}{ = 100, maximum surface area that can be wet (\%)}\cr\cr
\code{AK1inc}{ = 0.05, turns on thermal conductivity increase, the value being the increment by which AK1 is increased per iteration (W/mC)}\cr\cr
\code{AKMAX}{ = 2.8, maximum flesh conductivity (W/mC)}\cr\cr
\code{EXTinc}{ = 0.1, turns on changes in O2 extraction efficiency to simulate panting, the value being the increment by which AK1 is increased per iteration}\cr\cr

\strong{ General morphology:}\cr\cr
\code{ANDENS}{ = 1000, body density (kg/m3)}\cr\cr
\code{SUBQFAT}{ = 0, is subcutaneous fat present? (0 is no, 1 is yes)}\cr\cr
\code{FATPCT}{ = 20, \% body fat}\cr\cr
\code{GMULT}{ = GMREF, current ratio between long and short axis (-)}\cr\cr
\code{GMULTMAX}{ = GMREF, max possible ratio between long and short axis (-)}\cr\cr
\code{MAXPTVEN}{ = 0, maxium fraction of surface area that is ventral (dec \%)}\cr\cr
\code{AWING}{ = 0, area of wing, to do}\cr\cr
\code{PTCOND}{ = 0, \% of body area touching the substrate}\cr\cr

\strong{ Fur properties:}\cr\cr
\code{DHAIRD}{ = 30E-06, hair diameter, dorsal (m)}\cr\cr
\code{DHAIRV}{ = 30E-06, hair diameter, ventral (m)}\cr\cr
\code{LHAIRD}{ = 23.9E-03, hair length, dorsal (m)}\cr\cr
\code{LHAIRV}{ = 23.9E-03, hair length, dorsal (m)}\cr\cr
\code{RHOD}{ = 3000E+04, hair density, dorsal (1/m2)}\cr\cr
\code{RHOV}{ = 3000E+04, hair density, ventral (1/m2)}\cr\cr
\code{REFLD}{ = 0.2, fur reflectivity dorsal (dec \%)}\cr\cr
\code{REFLV}{ = 0.2, fur reflectivity ventral (dec \%)}\cr\cr

\strong{ Radiation exchange:}\cr\cr
\code{EMISAN}{ = 0.99, animal emissivity (-)}\cr\cr
\code{FATOBJ}{ = 0, configuration factor to nearby object}\cr\cr
\code{FABUSH}{ = 0, this is for veg below/around animal (at TALOC)}\cr\cr
\code{FGDREF}{ = 0.5, reference configuration factor to ground}\cr\cr
\code{FSKREF}{ = 0.5, configuration factor to sky}\cr\cr

\strong{ Nest properties:}\cr\cr
\code{NESTYP}{ = 0, # for nest calculations, to do)}\cr\cr
\code{RoNEST}{ = 0, # for nest calculations, to do}\cr\cr
\code{R1}{ = 0, # for nest calculations, to do}\cr\cr

\strong{ Physiology:}\cr\cr
\code{AK1}{ = 0.9, # initial thermal conductivity of flesh (0.412 - 2.8 W/mC)}\cr\cr
\code{AK2}{ = 0.230, # conductivity of fat (W/mC)}\cr\cr
\code{QBASAL}{ = (70 \* AMASS ^ 0.75) \* (4.185 / (24 \* 3.6)), # basal heat generation (W)}\cr\cr
\code{SKINW}{ = 0, # part of the skin surface that is wet (\%)}\cr\cr
\code{BAREVAP}{ = 0, # is evaporation partly from bare skin? (0 = no, 1 = yes, \% defined with PCTSKINEVAP)}\cr\cr
\code{PCTBAREVAP}{ = 2.8, maximum flesh conductivity (W/mC)}\cr\cr
\code{PCTEYES}{ = 0, # surface area made up by the eye (\%) - make zero if sleeping}\cr\cr
\code{DELTAR}{ = 0, # offset between air temperature and breath (°C)}\cr\cr
\code{RELXIT}{ = 100, # relative humidity of exhaled air, \%}\cr\cr
\code{TIMACT}{ = 1, # multiplier on metabolic rate for activity costs}\cr\cr
\code{RQ}{ = 0.80, # respiratory quotient (dec \%)}\cr\cr
\code{EXTREF}{ = 20, # O2 extraction efficiency (\%)}\cr\cr
\code{EXTMIN}{ = 6, # minimum O2 extraction efficiency (\%) to impose panting}\cr\cr
\code{Q10}{ = 2.5, # Q10 factor for adjusting BMR for TC}\cr\cr

\strong{ Initial conditions:}\cr\cr
\code{TS}{ = TC - 3, # initial skin temperature (°C)}\cr\cr
\code{TFA}{ = TA, # initial fur/air interface temperature (°C)}\cr\cr

\strong{Outputs:}
\code{TC}{ }\cr\cr
\code{TFA_D}{ }\cr\cr
\code{TFA_V}{ }\cr\cr
\code{TSKIN_D}{ }\cr\cr
\code{TSKIN_V}{ }\cr\cr
\code{QCONV_D}{ }\cr\cr
\code{QCONV_V}{ }\cr\cr
\code{QCOND_D}{ }\cr\cr
\code{QCOND_V}{ }\cr\cr
\code{QGENNET_D}{ }\cr\cr
\code{QGENNET_V}{ }\cr\cr
\code{QSEVAP_D}{ }\cr\cr
\code{QSEVAP_V}{ }\cr\cr
\code{QRAD_D}{ }\cr\cr
\code{QRAD_V}{ }\cr\cr
\code{QSLR_D}{ }\cr\cr
\code{QSLR_V}{ }\cr\cr
\code{QRSKY_D}{ }\cr\cr
\code{QRSKY_V}{ }\cr\cr
\code{QRBSH_D}{ }\cr\cr
\code{QRBSH_V}{ }\cr\cr
\code{QRVEG_D}{ }\cr\cr
\code{QRVEG_V}{ }\cr\cr
\code{QRGRD_D}{ }\cr\cr
\code{QRGRD_V}{ }\cr\cr
\code{NTRY_D}{ }\cr\cr
\code{NTRY_V}{ }\cr\cr
\code{SUCCESS_D}{ }\cr\cr
\code{SUCCESS_V}{ }\cr\cr
\code{RESPFN}{ }\cr\cr
\code{QRESP}{ }\cr\cr
\code{GEVAP}{ }\cr\cr
\code{PCTO2}{ }\cr\cr
\code{PCTN2}{ }\cr\cr
\code{PCTCO2}{ }\cr\cr
\code{RESPGEN}{ }\cr\cr
\code{O2STP}{ }\cr\cr
\code{O2MOL1}{ }\cr\cr
\code{N2MOL1}{ }\cr\cr
\code{AIRML1}{ }\cr\cr
\code{O2MOL2}{ }\cr\cr
\code{N2MOL2}{ }\cr\cr
\code{AIRML2}{ }\cr\cr
\code{AIRVOL}{ }\cr\cr
\code{GMULT}{ }\cr\cr
\code{SKINW}{ }\cr\cr
\code{SWEAT.G.H}{ }\cr\cr
\code{EVAP.G.H}{ }\cr\cr
\code{EXTREF}{ }\cr\cr
\code{AK}{ }\cr\cr
\code{TA}{ }\cr\cr
\code{TGRD}{ }\cr\cr
\code{TCONDSB}{ }\cr\cr
\code{TSKY}{ }\cr\cr
\code{VEL}{ }\cr\cr
\code{RH}{ }\cr\cr
\code{QSOLR}{ }\cr\cr
}
\examples{
library(NicheMapR)
dstart <- "02/01/2017"
dfinish <- "30/12/2017"
}
