% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_FACTOR.R
\name{F_FACTOR}
\alias{F_FACTOR}
\title{F_FACTOR}
\usage{
F_FACTOR(ASHADE, NITESHAD, QSOLR, FATOB, NESTYP, RoNEST, R1, FGDREF,
  FSKREF, AREASKIN, EMISAN)
}
\arguments{
\item{ASHADE}{A}

\item{NITESHAD}{A}

\item{QSOLR}{A}

\item{FATOB}{A}

\item{NESTYP}{A}

\item{RoNEST}{A}

\item{R1}{A}

\item{FGDREF}{A}

\item{FSKREF}{A}

\item{AREASKIN}{A}

\item{EMISAN}{A}
}
\description{
R wrapper for Fortran binary of F_FACTOR (endotherm model)
}
