% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gads.R
\name{gads.r}
\alias{gads.r}
\title{gads.r}
\usage{
gads.r(lat, lon, relhum, season)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{lon}{Longitude in decimal degrees}

\item{relhum}{Integer specifying relative humidity to use, 1:8 corresponds to 0,50,70,80,90,95,98,99 percent relative humidity}

\item{season}{Season to obtain data for, 0 = summer, 1 = winter}
}
\value{
optdep A vector of wavelength-specific optical depths
}
\description{
R function to extract data from the Fortran Global Aerosol Database software,
for a specific location, season and relative humidity
see http://www.mpimet.mpg.de/fileadmin/publikationen/Reports/MPI-Report_243.pdf
}
