% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOM.R
\name{GEOM}
\alias{GEOM}
\title{GEOM}
\usage{
GEOM(AMASS, ANDENS, FATPCT, POSTUR, ZFUR, SUBQFAT, GMULT, GMREF, DHARA,
  RHOARA, PTCOND, BIRD, MAMMAL, ORIENT)
}
\arguments{
\item{AMASS}{A}

\item{ANDENS}{A}

\item{FATPCT}{A}

\item{POSTUR}{A}

\item{ZFUR}{A}

\item{SUBQFAT}{A}

\item{GMULT}{A}

\item{GMREF}{A}

\item{DHARA}{A}

\item{RHOARA}{A}

\item{PTCOND}{A}

\item{BIRD}{A}

\item{MAMMAL}{A}

\item{ORIENT}{A}
}
\description{
R wrapper for Fortran binary of GEOM (endotherm model)
}
