% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOM_ecto.R
\encoding{UTF-8}
\name{GEOM_ecto}
\alias{GEOM_ecto}
\title{GEOM_ecto}
\usage{
GEOM_ecto(
  AMASS = 0.04,
  GEOMETRY = 3,
  SHP = c(1, 3, 2/3),
  CUSTOMGEOM = c(10.4713, 0.688, 0.425, 0.85, 3.798, 0.683, 0.694, 0.743),
  ANDENS = 1000,
  SKINW = 0.001,
  SKINT = 0,
  RINSUL = 0,
  PTCOND = 0.1,
  PMOUTH = 0.05,
  PANT = 1
)
}
\arguments{
\item{AMASS}{body mass, kg}

\item{GEOMETRY}{organism shape, 0-5, 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom}

\item{SHP}{vector of ratios of dimensions of long to short axes of linear dimensions}

\item{CUSTOMGEOM}{custom shape coefficients. Operates if shape=5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*mass^b, where AREA is in cm2 and mass is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}

\item{ANDENS}{animal density (kg/m3)}

\item{SKINW}{fraction of total surface area acting as a free water surface (fractional, 0-1)}

\item{SKINT}{fraction of total surface area in contact with another individual of the same temperature (fractional, 0-1)}

\item{RINSUL}{depth of insulation (m) (not yet used)}

\item{PTCOND}{fraction of total surface area in contact with the substrate (fractional, 0-1)}

\item{PMOUTH}{fraction of total surface area that is 'wet' mouth (fractional, 0-1)}

\item{PANT}{multiplier on breathing rate, for respiratory water loss via panting (used here to open mouth if > 1)}
}
\description{
R version of Fortran GEOM.f (ectotherm model) for calculating biophysically relevant object surfaces, volumes and linear dimensions.
}
