% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOM.R
\encoding{UTF-8}
\name{GEOM_ENDO}
\alias{GEOM_ENDO}
\title{GEOM_ENDO}
\usage{
GEOM_ENDO(
  AMASS,
  ANDENS,
  FATPCT,
  SHAPE,
  ZFUR,
  SUBQFAT,
  SHAPE_B,
  SHAPE_C,
  DHARA,
  RHOARA,
  PCOND,
  SAMODE,
  ORIENT,
  ZEN
)
}
\arguments{
\item{AMASS}{mass, kg}

\item{ANDENS}{volume, kg/m3}

\item{FATPCT}{fat percentage, \%}

\item{SHAPE}{shape, -}

\item{ZFUR}{fur depth, m}

\item{SUBQFAT}{subcutaneous fat presence, -}

\item{SHAPE_B}{multiplier on long axis for shape change, -}

\item{SHAPE_C}{multiplier on width relative to height for plate shape change, -}

\item{DHARA}{hair diameter, m}

\item{RHOARA}{hair density, 1/m2}

\item{PCOND}{fraction of surface area conducting to substrate, -}

\item{SAMODE}{surface area mode, -}

\item{ORIENT}{orientation to solar, -}

\item{ZEN}{zenith angle, degrees}
}
\description{
R wrapper for Fortran binary of GEOM_ENDO (endotherm model)
}
