% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.global.climate.R
\name{get.global.climate}
\alias{get.global.climate}
\title{Downloads the global climate, soil moisture and elevation datasets
required to run function micro_global}
\usage{
get.global.climate(folder)
}
\arguments{
\item{folder}{Path to the folder you want to install the global climate data in}
}
\description{
the files are:
global
derived from "New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data
set of surface climate over global land areas. Climate Research 21:1-25"
It also optionally uses a global monthly soil moisture estimate from NOAA CPC Soil Moisture http://140.172.38.100/psd/thredds/catalog/Datasets/cpcsoil/catalog.html
Aerosol attenuation can also be computed based on the Global Aerosol Data Set (GADS)
}

