% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.longlat.R
\encoding{UTF-8}
\name{get.longlat}
\alias{get.longlat}
\title{Wrapper for google_geocode from the googleway package.}
\usage{
get.longlat(place = "Alice Springs, Australia", key = 'your key')
}
\arguments{
\item{place}{String describing the location, e.g. "Alice Springs, Australia"}

\item{key}{Your API key}
}
\value{
longlat The longitude and latitude in decimal degrees
}
\description{
A Google Maps API Key is required for this function, which can be obtained at https://cloud.google.com/maps-platform/ (credit card details need to be provided but it can be used for free for a certain limit of queries)
}
