% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clo.R
\encoding{UTF-8}
\name{get_clo}
\alias{get_clo}
\title{get_clo - auxiliary function for HomoTherm}
\usage{
get_clo(HomoTherm.out, INSDEPDs = c(1e-02, rep(4.1e-03, 3)), INSDEPVs = c(1e-09, rep(4.1e-03, 3)))
}
\arguments{
\item{indata}{the HomoTherm output list}

\item{INSDEPDs}{= c(1e-02, rep(6e-03, 3)), clothing depth, dorsal (m)}

\item{INSDEPVs}{= c(1e-09, rep(6e-03, 3)), clothing depth, ventral (m)}

\item{AREAFRACs}{= c(0.08291887, 0.32698460, 0.11025155, 0.18479669), fractional areas (-)}
}
\description{
This function takes the output from the HomoTherm model and computes the
surface area-weighted insulation, reported back in 'clo' units.
}
