% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pressure.R
\name{get_pressure}
\alias{get_pressure}
\title{Atmospheric pressure from reference elevation}
\usage{
get_pressure(h = 1000)
}
\arguments{
\item{h_ref}{= 0, reference elevation, m}

\item{P_ref}{= 101325, static pressure at sea level, Pa}

\item{L_ref}{= -0.0065, reference temperature lapse rate, K/m}

\item{T_ref}{= 288, reference temperature (at sea level), K}

\item{g_0}{= 9.80665, gravitational acceleration constant, m/s^2}

\item{M}{= 0.0289644, molar mass of air, kg/mol}
}
\value{
pressure at focal height, Pa
}
\description{
Function to get atmospheric pressure at an elevation of interest (h), given
reference level pressure, temperature, lapse rate and molar mass of air
}
