% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terra.R
\name{get_terra}
\alias{get_terra}
\title{Function to extract point data from TerraClimate}
\usage{
get_terra(
  scenario = 0,
  x = c(-5.3, 50.13),
  ystart = 1985,
  yfinish = 2015,
 
    source = "http://thredds.northwestknowledge.net:8080/thredds/dodsC/TERRACLIMATE_ALL/data"
)
}
\arguments{
\item{scenario}{= 0, climate scenario, either 0 (historical), 2 (plus 2 deg C) or 4 (plus 4 deg C)}

\item{x}{= c(-5.3, 50.13), location expressed as longitude, latitude, decimal degrees}

\item{ystart}{= start year (min is 1958 for historical, 1985 for climate change)}

\item{yfinish}{= finish year (max is 'last year' for historical, 2015 for climate change)}

\item{source}{= "http://thredds.northwestknowledge.net:8080/thredds/dodsC/TERRACLIMATE_ALL/data", location where data is stored - defaults to opendap web location but can be retrieved locally if desired}
}
\value{
TMINN, minimum daily 2m air temperature, deg C

TMAXX, maximum daily 2m air temperature, deg C

RAINFALL, daily total precipitation, mm

VPD, daily vapour pressure deficit, kPa

SRAD, daily solar radiation, W/m2

SoilMoist, volumetric soil moisture, m3/m3

WIND, daily 10m wind speed, m/s
}
\description{
Extracts variables needed for microclimate modelling from the monthly TerraClimate database
described in "Abatzoglou, J. T., Dobrowski, S. Z., Parks, S. A., & Hegewisch, K. C. (2018). TerraClimate,
a high-resolution global dataset of monthly climate and climatic water balance from 1958–2015.
Scientific Data, 5(1), 170191. https://doi.org/10.1038/sdata.2017.191"
This dataset includes climate change scenarios for +2 and +4 deg C warming (via parameter 'scenario')
Michael Kearney Feb 2022
}
