% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPROP.R
\name{IRPROP}
\alias{IRPROP}
\title{IRPROP}
\usage{
IRPROP(TA, SHAPE_B_MAX, SHAP_B_REF, SHAPE_B, DHAIRD, DHAIRV, LHAIRD,
  LHAIRV, ZFURD, ZFURV, RHOD, RHOV, REFLD, REFLV, MAXPTVEN, ZFURCOMP)
}
\arguments{
\item{TA}{A}

\item{SHAPE_B_MAX}{A}

\item{SHAP_B_REF}{A}

\item{SHAPE_B}{A}

\item{DHAIRD}{A}

\item{DHAIRV}{A}

\item{LHAIRD}{A}

\item{LHAIRV}{A}

\item{ZFURD}{A}

\item{ZFURV}{A}

\item{RHOD}{A}

\item{RHOV}{A}

\item{REFLD}{A}

\item{REFLV}{A}

\item{MAXPTVEN}{A}

\item{ZFURCOMP}{A}
}
\description{
R wrapper for Fortran binary of IRPROP (endotherm model)
}
