% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfix.R
\encoding{UTF-8}
\name{leapfix}
\alias{leapfix}
\title{Function to assist with interpolated data in leap years}
\usage{
leapfix(indata, yearlist, mult = 1)
}
\arguments{
\item{indata}{data needing to be interpolated}

\item{yearlist}{list of years to do}

\item{mult}{how many elements per day? (1 = daily, 24 = hourly)}
}
\description{
Function to assist with interpolated data in leap years
}
