% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_clima.R
\name{micro_clima}
\alias{micro_clima}
\title{NCEP implementation of the microclimate model, with microclima computing hourly forcing.}
\usage{
micro_ncep(loc = 'Galapagos', dstart = "01-01-2017", dfinish = "31-12-2017",
REFL = 0.15, slope = 0, aspect = 0, DEP = c(0, 2.5,  5,  10,  15,  20,  30,  50,  100,  200), minshade = 0, maxshade = 90,
Usrhyt = 0.01, ...)
}
\arguments{
\item{loc}{Either a longitude and latitude (decimal degrees) or a place name to search for on Google Earth}

\item{dstart}{First day to run, date in format "d-m-Y" e.g. "01-01-2016"}

\item{dfinish}{Last day to run, date in format "d-m-Y" e.g. "31-12-2016"}

\item{REFL}{Soil solar reflectance, decimal \%}

\item{elev}{Elevation, if to be user specified (m)}

\item{slope}{Slope in degrees}

\item{aspect}{Aspect in degrees (0 = north)}

\item{DEP}{Soil depths at which calculations are to be made (cm), must be 10 values starting from 0, and more closely spaced near the surface}

\item{minshade}{Minimum shade level to use (\%)}

\item{maxshade}{Maximum shade level to us (\%)}

\item{Usrhyt}{Local height (m) at which air temperature, wind speed and humidity are to be computed for organism of interest}

\item{...}{Additional arguments, see Details}
}
\value{
metout The above ground micrometeorological conditions under the minimum specified shade

shadmet The above ground micrometeorological conditions under the maximum specified shade

soil Hourly predictions of the soil temperatures under the minimum specified shade

shadsoil Hourly predictions of the soil temperatures under the maximum specified shade

soilmoist Hourly predictions of the soil moisture under the minimum specified shade

shadmoist Hourly predictions of the soil moisture under the maximum specified shade

soilpot Hourly predictions of the soil water potential under the minimum specified shade

shadpot Hourly predictions of the soil water potential under the maximum specified shade

humid Hourly predictions of the soil humidity under the minimum specified shade

shadhumid Hourly predictions of the soil humidity under the maximum specified shade

plant Hourly predictions of plant transpiration, leaf water potential and root water potential under the minimum specified shade

shadplant Hourly predictions of plant transpiration, leaf water potential and root water potential under the maximum specified shade

sunsnow Hourly predictions of snow temperature under the minimum specified shade

shadsnow Hourly predictions snow temperature under the maximum specified shade
}
\description{
An implementation of the NicheMapR microclimate model that uses the RNCEP package daily weather database https://sites.google.com/site/michaelukemp/rncep, and specifically uses the following variables: air.2m, prate.sfc, shum.2m, pres.sfc, tcdc.eatm, uwnd.10m, vwnd.10m, dswrf.sfc. At the moment uses the same DEM from the CRU global climate data set.
}
\examples{
library(NicheMapR)
dstart <- "01/01/1976"
dfinish <- "31/12/1978"
micro<-micro_ncep(loc = 'Galapagos', runshade = 0, soilgrids = 0, dstart = dstart, dfinish = dfinish, reanalysis = FALSE)

metout<-as.data.frame(micro$metout) # above ground microclimatic conditions, min shade
soil<-as.data.frame(micro$soil) # soil temperatures, minimum shade
soilmoist<-as.data.frame(micro$soilmoist) # soil temperatures, minimum shade

# append dates
tzone<-paste("Etc/GMT+",0,sep="")
dates<-seq(as.POSIXct(dstart, format="\%d/\%m/\%Y",tz=tzone)-3600*12, as.POSIXct(dfinish, format="\%d/\%m/\%Y",tz=tzone)+3600*11, by="hours")

metout <- cbind(dates,metout)
soil <- cbind(dates,soil)
soilmoist <- cbind(dates, soilmoist)
minshade<-micro$minshade

# plotting above-ground conditions in minimum shade
with(metout,{plot(TALOC ~ dates,xlab = "Date and Time", ylab = "Temperature (°C)"
, type = "l",main=paste("air and sky temperature, ",minshade,"\% shade",sep=""), ylim = c(-20, 60))})
with(metout,{points(TAREF ~ dates,xlab = "Date and Time", ylab = "Temperature (°C)"
, type = "l",lty=2,col='blue')})
with(metout,{points(TSKYC ~ dates,xlab = "Date and Time", ylab = "Temperature (°C)"
,  type = "l",col='light blue',main=paste("sky temperature, ",minshade,"\% shade",sep=""))})
with(metout,{plot(RHLOC ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)"
, type = "l",ylim=c(0,100),main=paste("humidity, ",minshade,"\% shade",sep=""))})
with(metout,{points(RH ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)"
, type = "l",col='blue',lty=2,ylim=c(0,100))})
with(metout,{plot(VREF ~ dates,xlab = "Date and Time", ylab = "Wind Speed (m/s)"
,  type = "l",main="wind speed",ylim = c(0, 15))})
with(metout,{points(VLOC ~ dates,xlab = "Date and Time", ylab = "Wind Speed (m/s)"
,  type = "l",lty=2,col='blue')})
with(metout,{plot(SOLR ~ dates,xlab = "Date and Time", ylab = "Solar Radiation (W/m2)"
,  type = "l",main="solar radiation")})
with(metout,{plot(SNOWDEP ~ dates,xlab = "Date and Time", ylab = "Snow Depth (cm)"
,  type = "l",main="snow depth")})

# plotting soil temperature
for(i in 1:10){
 if(i==1){
   plot(soil[,i+3]~soil[,1],xlab = "Date and Time", ylab = "Soil Temperature (°C)"
   ,col=i,type = "l",main=paste("soil temperature ",minshade,"\% shade",sep=""))
 }else{
   points(soil[,i+3]~soil[,1],xlab = "Date and Time", ylab = "Soil Temperature
    (°C)",col=i,type = "l")
 }
}

# plotting soil moisture
for(i in 1:10){
 if(i==1){
   plot(soilmoist[,i+3]*100~soilmoist[,1],xlab = "Date and Time", ylab = "Soil Moisture (\% volumetric)"
   ,col=i,type = "l",main=paste("soil moisture ",minshade,"\% shade",sep=""))
 }else{
   points(soilmoist[,i+3]*100~soilmoist[,1],xlab = "Date and Time", ylab = "Soil Moisture
    (\%)",col=i,type = "l")
 }
}
}
