% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_global.R
\name{micro_global}
\alias{micro_global}
\title{Global implementation of the microclimate model.}
\usage{
micro_global(loc, timeinterval, nyears, soiltype, REFL, slope, aspect, DEP, minshade, maxshade, Usrhyt, ...)
}
\arguments{
\item{loc}{Either a longitude and latitude (decimal degrees) or a place name to search for on Google Earth}

\item{timeinterval}{The number of time intervals to generate predictions for over a year (must be 12 <= x <=365)}

\item{nyears}{The number of years to run}

\item{soiltype}{Soil type: Rock = 0, sand = 1, loamy sand = 2, sandy loam = 3, loam = 4, silt loam = 5, sandy clay loam = 6, clay loam = 7, silt clay loam = 8, sandy clay = 9, silty clay = 10, clay = 11, based on Campbell and Norman 1990 Table 9.1.}

\item{REFL}{Soil solar reflectance, decimal percent}

\item{slope}{Slope in degrees}

\item{aspect}{Aspect in degrees (0 = north)}

\item{DEP}{Soil depths at which calculations are to be made (cm), must be 10 values starting from 0, and more closely spaced near the surface}

\item{minshade}{Minimum shade level to use (percent)}

\item{maxshade}{Maximum shade level to us (percent)}

\item{Usrhyt}{Local height (cm) at which air temperature, wind speed and humidity are to be computed for organism of interest}
}
\value{
metout The above ground micrometeorological conditions under the minimum specified shade

shadmet The above ground micrometeorological conditions under the maximum specified shade

soil Hourly predictions of the soil temperatures under the minimum specified shade

shadsoil Hourly predictions of the soil temperatures under the maximum specified shade

soilmoist Hourly predictions of the soil moisture under the minimum specified shade

shadmoist Hourly predictions of the soil moisture under the maximum specified shade

soilpot Hourly predictions of the soil water potential under the minimum specified shade

shadpot Hourly predictions of the soil water potential under the maximum specified shade

humid Hourly predictions of the soil humidity under the minimum specified shade

shadhumid Hourly predictions of the soil humidity under the maximum specified shade
}
\description{
An implementation of the Niche Mapper microclimate model that uses the global climate database
derived from "New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data
set of surface climate over global land areas. Climate Research 21:1-25"
It also optionally uses a global monthly soil moisture estimate from NOAA CPC Soil Moisture http://140.172.38.100/psd/thredds/catalog/Datasets/cpcsoil/catalog.html
Aerosol attenuation can also be computed based on the Global Aerosol Data Set (GADS)
Koepke, P., M. Hess, I. Schult, and E. P. Shettle. 1997. Global Aerosol Data Set. Max-Planck-Institut for Meteorologie, Hamburg
by choosing the option 'rungads<-1'
}
\details{
metout/shadmet variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 TIME - time of day (mins)
\item 3 TALOC - air temperature (deg C) at local height (specified by 'Usrhyt' variable)
\item 4 TAREF - air temperature (deg C) at reference height (1.2m)
\item 5 RHLOC - relative humidity (percent) at local height (specified by 'Usrhyt' variable)
\item 6 RH  - relative humidity (percent) at reference height (1.2m)
\item 7 VLOC - wind speed (m/s) at local height (specified by 'Usrhyt' variable)
\item 8 VREF - wind speed (m/s) at reference height (1.2m)
\item 9 SNOWMELT - snowmelt (mm)
\item 10 POOLDEP - water pooling on surface (mm)
\item 11 PCTWET - soil surface wetness (percent)
\item 12 ZEN - zenith angle of sun (degrees - 90 = below the horizon)
\item 13 SOLR - solar radiation (W/m2)
\item 14 TSKYC - sky radiant temperature (deg C)
\item 15 DEW - dew presence (0 or 1)
\item 16 FROST - frost presence (0 or 1)
\item 17 SNOWFALL - snow predicted to have fallen (mm)
\item 18 SNOWDEP - predicted snow depth (cm)
}
soil and shadsoil variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 TIME - time of day (mins)
\item 3-12 WC0cm ... - soil temperatures at each of the 10 specified depths

if soil moisture model is run i.e. parameter runmoist = 1\cr

soilmoist and shadmoist variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 TIME - time of day (mins)
}
soilpot and shadpot variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 TIME - time of day (mins)
\item 3-12 PT0cm ... - soil water potential (J/kg = kpa = bar/100) at each of the 10 specified depths
}

humid and shadhumid variables:
\itemize{
\item  1 JULDAY - day of year
\item  2 TIME - time of day (mins)
\item  3-12 RH0cm ... - soil relative humidity (decimal percent), at each of the 10 specified depths
}
}
}
\examples{
micro<-micro_global() # run the model with default location and settings

metout<-as.data.frame(micro$metout) # above ground microclimatic conditions, min shade
shadmet<-as.data.frame(micro$shadmet) # above ground microclimatic conditions, max shade
soil<-as.data.frame(micro$soil) # soil temperatures, minimum shade
shadsoil<-as.data.frame(micro$shadsoil) # soil temperatures, maximum shade

# append dates
days<-rep(seq(1,12),24)
days<-days[order(days)]
dates<-days+metout$TIME/60/24-1 # dates for hourly output
dates2<-seq(1,12,1) # dates for daily output

plotmetout<-cbind(dates,metout)
plotsoil<-cbind(dates,soil)
plotshadmet<-cbind(dates,shadmet)
plotshadsoil<-cbind(dates,shadsoil)

minshade<-micro$minshade
maxshade<-micro$maxshade

# plotting above-ground conditions in minimum shade
with(plotmetout,{plot(TALOC ~ dates,xlab = "Date and Time", ylab = "Air Temperature (deg C)", type = "l",main=paste("air temperature, ",minshade,"\% shade",sep=""))})
with(plotmetout,{points(TAREF ~ dates,xlab = "Date and Time", ylab = "Air Temperature (deg C)", type = "l",lty=2,col='blue')})
with(plotmetout,{plot(RHLOC ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)", type = "l",ylim=c(0,100),main=paste("humidity, ",minshade,"\% shade",sep=""))})
with(plotmetout,{points(RH ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)", type = "l",col='blue',lty=2,ylim=c(0,100))})
with(plotmetout,{plot(TSKYC ~ dates,xlab = "Date and Time", ylab = "Sky Temperature (deg C)",  type = "l",main=paste("sky temperature, ",minshade,"\% shade",sep=""))})
with(plotmetout,{plot(VREF ~ dates,xlab = "Date and Time", ylab = "Wind Speed (m/s)",  type = "l",main="wind speed")})
with(plotmetout,{points(VLOC ~ dates,xlab = "Date and Time", ylab = "Wind Speed (m/s)",  type = "l",lty=2,col='blue')})
with(plotmetout,{plot(ZEN ~ dates,xlab = "Date and Time", ylab = "Zenith Angle of Sun (deg)",  type = "l",main="solar angle, sun")})
with(plotmetout,{plot(SOLR ~ dates,xlab = "Date and Time", ylab = "Solar Radiation (W/m2)",  type = "l",main="solar radiation")})

# plotting soil temperature for minimum shade
for(i in 1:10){
 if(i==1){
   plot(plotsoil[,i+3]~plotsoil[,1],xlab = "Date and Time", ylab = "Soil Temperature (deg C)",col=i,type = "l",main=paste("soil temperature ",minshade,"\% shade",sep=""))
 }else{
   points(plotsoil[,i+3]~plotsoil[,1],xlab = "Date and Time", ylab = "Soil Temperature (deg C)",col=i,type = "l")
 }
}

# plotting above-ground conditions in maximum shade
with(plotshadmet,{plot(TALOC ~ dates,xlab = "Date and Time", ylab = "Air Temperature (deg C)", type = "l",main="air temperature, sun")})
with(plotshadmet,{points(TAREF ~ dates,xlab = "Date and Time", ylab = "Air Temperature (deg C)", type = "l",lty=2,col='blue')})
with(plotshadmet,{plot(RHLOC ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)", type = "l",ylim=c(0,100),main="humidity, shade")})
with(plotshadmet,{points(RH ~ dates,xlab = "Date and Time", ylab = "Relative Humidity (\%)", type = "l",col='blue',lty=2,ylim=c(0,100))})
with(plotshadmet,{plot(TSKYC ~ dates,xlab = "Date and Time", ylab = "Sky Temperature (deg C)",  type = "l",main="sky temperature, shade")})

# plotting soil temperature for maximum shade
for(i in 1:10){
 if(i==1){
   plot(plotshadsoil[,i+3]~plotshadsoil[,1],xlab = "Date and Time", ylab = "Soil Temperature (deg C)",col=i,type = "l",main=paste("soil temperature ",maxshade,"\% shade",sep=""))
 }else{
   points(plotshadsoil[,i+3]~plotshadsoil[,1],xlab = "Date and Time", ylab = "Soil Temperature (deg C)",col=i,type = "l")
 }
}
}

