% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microrun.R
\name{microclimate}
\alias{microclimate}
\title{microclimate model}
\usage{
microclimate(micro)
}
\arguments{
\item{micro}{A vector of input variables for the microclimate model}
}
\value{
metout The above ground micrometeorological conditions under the minimum specified shade

shadmet The above ground micrometeorological conditions under the maximum specified shade

soil Hourly predictions of the soil temperatures under the minimum specified shade

shadsoil Hourly predictions of the soil temperatures under the maximum specified shade

soilmoist Hourly predictions of the soil moisture under the minimum specified shade

shadmoist Hourly predictions of the soil moisture under the maximum specified shade

soilpot Hourly predictions of the soil water potential under the minimum specified shade

shadpot Hourly predictions of the soil water potential under the maximum specified shade

humid Hourly predictions of the soil humidity under the minimum specified shade

shadhumid Hourly predictions of the soil humidity under the maximum specified shade

plant Hourly predictions of plant variables under the minimum specified shade

shadplant Hourly predictions of plant variables under the maximum specified shade

sunsnow Hourly predictions of the snow temperature under the minimum specified shade

shdsnow Hourly predictions of the snow temperature under the maximum specified shade

tcond Hourly predictions of the soil thermal conductivity under the minimum specified shade

shadtcond Hourly predictions of the soil thermal conductivity under the maximum specified shade

specheat Hourly predictions of the soil specific heat capacity under the minimum specified shade

shadspecheat Hourly predictions of soil specific heat capacity under the maximum specified shade

densit Hourly predictions of the soil density under the minimum specified shade

shaddensit Hourly predictions of the soil density under the maximum specified shade
}
\description{
R wrapper for Fortran binary of Niche Mapper microclimate model
}
