% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NESTCALC.R
\encoding{UTF-8}
\name{NESTCALC}
\alias{NESTCALC}
\title{NESTCALC}
\usage{
NESTCALC(NESTCALC.input)
}
\arguments{
\item{NESTCALC.input}{description}
}
\description{
R wrapper for Fortran binary of NESTCALC (endotherm model)
}
