% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onelump.R
\name{onelump}
\alias{onelump}
\title{One-lump Transient Heat Budget for Constant Environment}
\usage{
onelump(t, Tc_init, thresh, AMASS, lometry, Tair, Trad, vel, Qsol, Zen, ...)
}
\arguments{
\item{t}{= seq(1,3600,60), time intervals (s) at which output is required}

\item{Tc_init}{= 5, initial temperature (degrees C) Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see details)}

\item{thresh}{= 29, threshold temperature (degrees C) at which summary statistics are wanted}

\item{AMASS}{= 500, animal mass (g)}

\item{lometry}{= 2, Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see parameter 'customallom')}

\item{Tair}{= 30, air temperature (degrees C)}

\item{Trad}{= 30, radiant temperature (degrees C), averaging ground and sky}

\item{vel}{= 0.1, wind speed (m/s)}

\item{Qsol}{= 500, solar radiation (W/m2)}

\item{Zen}{= 20, zenith angle of sun (90 is below horizon), degrees}

\item{Flshcond}{= 0.5, Thermal conductivity of flesh (W/mK, range: 0.412-2.8)}

\item{q}{= 0, metabolic heat production rate W/m3}

\item{Spheat}{= 0.85, Specific heat of flesh J/(kg-K)}

\item{EMISAN}{= 0.95, Emissivity of animal (0-1)}

\item{ABS}{= 0.85, solar absorptivity, decimal percent}

\item{customallom}{= c(10.4713,.688,0.425,0.85,3.798,.683,0.694,.743), Custom allometry coefficients. Operates if lometry=5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*mass^b, where AREA is in cm2 and mass is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}

\item{shape_a}{= 1., Proportionality factor (-) for going from volume to area, keep this 1 (redundant parameter that should be removed)}

\item{shape_b}{= 3, Proportionality factor (-) for going from volume to area, represents ratio of width:height for a plate, length:diameter for cylinder, b axis:a axis for ellipsoid}

\item{shape_c}{= 0.6666666667, Proportionality factor (-) for going from volume to area, represents ratio of length:height for a plate, c axis:a axis for ellipsoid}

\item{posture}{= 'n' pointing normal 'n', parallel 'p' to the sun's rays, or 'b' in between?}

\item{FATOSK}{= 0.4, Configuration factor to sky (-) for infrared calculations}

\item{FATOSB}{= 0.4, Configuration factor to subsrate for infrared calculations}

\item{sub_reflect}{= 0.2, substrate solar reflectivity, decimal percent}

\item{PCTDIF}{= 0.1, proportion of solar energy that is diffuse (rather than direct beam)}

\item{press}{= 101325, air pressure (Pa)}

\item{VTMIN}{= 24, Voluntary thermal minimum, degrees C (lower body temperature for foraging)}
}
\value{
Tc Core temperature (deg C)

Tcf Final (steady state) temperature (deg C), if conditions remained constant indefinately

tau Time constant (s)

dTc Rate of change of core temperature (deg C/s)

timethresh Time to reach specified threshold body temperature (s)
}
\description{
Transient, 'one-lump', heat budget for computing rate of change of temperature
under a constant environment
Michael Kearney & Warren Porter developed this R function in July 2014.
}
\examples{
# compare heating rates of a variety of different-sized objects

t=seq(1,3600*2,60) # times (in seconds) to report back values - every minute for 2 hrs

masses=c(1,50,100,500,1000) # masses of 5 objects to consider

# loop through masses and plot warming curves for the 5 objects, starting at 5 degrees C, no solar
for(i in 1:5){
Tbs<-onelump(t=t, Tc_init=5, thresh=29, AMASS=masses[i], lometry=2, Tair=30, Trad=30, vel=0.1, Qsol=0, Zen=20)
 if(i==1){
  plot(Tbs$Tc~t,type='l',col=i,ylim=c(0,50),ylab='Temperature, deg C',xlab='time, s', main='30 degrees air, no solar radiation')
 }else{
  points(Tbs$Tc~t,type='l',col=i)
 }
}

# repeat analysis, but with sun on the objects
for(i in 1:5){
Tbs<-onelump(t=t, Tc_init=5, thresh=29, AMASS=masses[i], lometry=2, Tair=30, Trad=30, vel=0.1, Qsol=1000, Zen=20)
 if(i==1){
  plot(Tbs$Tc~t,type='l',col=i,ylim=c(0,50),ylab='Temperature, deg C',xlab='time, s', main='30 degrees air, strong solar radiation field')
 }else{
  points(Tbs$Tc~t,type='l',col=i)
 }
}
}

