% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADOUT_ecto.R
\encoding{UTF-8}
\name{RADOUT_ecto}
\alias{RADOUT_ecto}
\title{RADOUT_ecto}
\usage{
RADOUT_ecto(
  TSKIN = 25.1,
  ATOT = 0.01325006,
  AV = 0.001325006,
  AT = 0,
  FATOSK = 0.4,
  FATOSB = 0.4,
  EMISAN = 0.95
)
}
\arguments{
\item{TSKIN}{skin temperature (°C)}

\item{ATOT}{total body surface area (m2)}

\item{AV}{ventral surface area (m2)}

\item{AT}{body surface area contacting another organism of same temperature (m2)}

\item{FATOSK}{configuration factor to sky (-)}

\item{FATOSB}{configuration factor to substrate (-)}

\item{EMISAN}{emissivity of animal (fractional, 0-1)}
}
\description{
R version of Fortran RADOUT.f (ectotherm model) for calculating outgoing long-wave radiation.
}
