% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESP_ecto.R
\encoding{UTF-8}
\name{RESP_ecto}
\alias{RESP_ecto}
\title{RESP_ecto}
\usage{
RESP_ecto(
  XTRY = 23,
  AMASS = 0.04,
  TC = 25,
  QMETAB = 0.01241022,
  EXTREF = 20,
  PANT = 1,
  RQ = 0.8,
  TA = 20,
  RELHUM = 50,
  BP = 101325,
  O2GAS = 20.95,
  CO2GAS = 0.03,
  N2GAS = 79.02,
  DELTAR = 0.1
)
}
\arguments{
\item{XTRY}{current guess of lung temperature (째C)}

\item{AMASS}{body mass, kg}

\item{TC}{current guess of core body temperature (째C)}

\item{QMETAB}{metabolic rate calculated with MET_ecto (W)}

\item{EXTREF}{oxygen extraction efficiency (\%)}

\item{PANT}{multiplier on breathing rate, for respiratory water loss via panting}

\item{RQ}{respiratory quotient (fractional, 0-1)}

\item{TA}{air temperature (째C)}

\item{RELHUM}{relative humidity (\%)}

\item{BP}{air pressure (Pa)}

\item{O2GAS}{oxygen concentration of atmosphere (\%)}

\item{CO2GAS}{carbon dioxide concentration of atmosphere (\%)}

\item{N2GAS}{nitrogen concentration of atmosphere (\%)}

\item{DELTAR}{temperature difference (째C) between expired and inspired air}
}
\description{
R version of Fortran RESP.f (ectotherm model) for calculating respiratory heat and water exchange.
}
