% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEVAP.R
\name{SEVAP}
\alias{SEVAP}
\title{SEVAP}
\usage{
SEVAP(BP, TA, RELHUM, VEL, TC, TSURF, ALT, SKINW, FLYHR, CONVSK, HD,
  HDFREE, PCTBAREVAP, PCTEYES, ZFUR, FURWET, TFA, CONVAR)
}
\arguments{
\item{BP}{A}

\item{TA}{A}

\item{RELHUM}{A}

\item{VEL}{A}

\item{TC}{A}

\item{TSURF}{A}

\item{ALT}{A}

\item{SKINW}{A}

\item{FLYHR}{A}

\item{CONVSK}{A}

\item{HD}{A}

\item{HDFREE}{A}

\item{PCTBAREVAP}{A}

\item{PCTEYES}{A}

\item{ZFUR}{A}

\item{FURWET}{A}

\item{TFA}{A}

\item{CONVAR}{A}
}
\description{
R wrapper for Fortran binary of SEVAP (endotherm model)
}
