% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEVAP_ecto.R
\encoding{UTF-8}
\name{SEVAP_ecto}
\alias{SEVAP_ecto}
\title{SEVAP_ecto}
\usage{
SEVAP_ecto(
  TC = 25,
  TSKIN = 25.1,
  GEVAP = 1.177235e-09,
  PSI_BODY = -7.07 * 100,
  SKINW = 0.1,
  AEFF = 1.192505e-05,
  ATOT = 0.01325006,
  HD = 0.02522706,
  PEYES = 0.03/100,
  TA = 20,
  RELHUM = 50,
  VEL = 0.1,
  BP = 101325
)
}
\arguments{
\item{TC}{core body temperature (°C)}

\item{TSKIN}{skin temperature (°C)}

\item{GEVAP}{respiratory water loss rate (just used for getting total water loss) (kg/s)}

\item{PSI_BODY}{water potential of body (J/kg) - affects skin humidity for water vapour exchange}

\item{SKINW}{fraction of total surface area acting as a free water surface (fractional, 0-1)}

\item{AEFF}{effective area acting as a free-water exchanger, drives cutaneous evaporation (m2)}

\item{ATOT}{total body surface area (m2)}

\item{HD}{mass transfer coefficient calculated with CONV_ecto (m/s)}

\item{PEYES}{proportion of total surface area that is 'wet' eye (fractional, 0-1)}

\item{TA}{air temperature (°C)}

\item{RELHUM}{relative humidity (\%)}

\item{VEL}{wind speed (m/s)}

\item{BP}{air pressure (Pa)}
}
\description{
R version of Fortran SEVAP.f (ectotherm model)
}
