% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULSOL.R
\encoding{UTF-8}
\name{SIMULSOL}
\alias{SIMULSOL}
\title{SIMULSOL}
\usage{
SIMULSOL(DIFTOL, IPT, FURVARS, GEOMVARS, ENVVARS, TRAITS, TFA, SKINW,
  TSKIN, results)
}
\arguments{
\item{DIFTOL}{A}

\item{IPT}{A}

\item{FURVARS}{A}

\item{GEOMVARS}{A}

\item{ENVVARS}{A}

\item{TRAITS}{A}

\item{TFA}{A}

\item{SKINW}{A}

\item{TSKIN}{A}
}
\description{
R wrapper for Fortran binary of SIMULSOL (endotherm model)
}
