% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR.R
\name{SOLAR}
\alias{SOLAR}
\title{SOLAR}
\usage{
SOLAR(AREATOTL, ABSAND, ABSANV, ABSSB, ASILN, PCTDIF, QNORM, SHADE, QSOLR,
  FASKY, FATOBJ, FAVEG)
}
\arguments{
\item{AREATOTL}{A}

\item{ABSAND}{A}

\item{ABSANV}{A}

\item{ABSSB}{A}

\item{ASILN}{A}

\item{PCTDIF}{A}

\item{QNORM}{A}

\item{SHADE}{A}

\item{QSOLR}{A}

\item{FASKY}{A}

\item{FATOBJ}{A}

\item{FAVEG}{A}
}
\description{
R wrapper for Fortran binary of SOLAR (endotherm model)
}
