% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR_ecto.R
\encoding{UTF-8}
\name{SOLAR_ecto}
\alias{SOLAR_ecto}
\title{SOLAR_ecto}
\usage{
SOLAR_ecto(
  ATOT = 0.01325006,
  ASIL = 0.004718043,
  AV = 0.001325006,
  AT = 0,
  ABSAN = 0.85,
  ABSSB = 0.8,
  FATOSK = 0.4,
  FATOSB = 0.4,
  FATOBJ = 0,
  ZEN = 20 * pi/180,
  QSOLR = 1000,
  PDIF = 0.1,
  SHADE = 0,
  postur = 1
)
}
\arguments{
\item{ATOT}{surface area for solar exchange, m2}

\item{ASIL}{silouette area for solar exchange, m2}

\item{AV}{ventral area for solar exchange with substrate, m2}

\item{AT}{total surface area touching another animal, m2}

\item{ABSAN}{solar absorptivity of animal (fractional, 0-1)}

\item{ABSSB}{solar absorptivity of substrate (fractional, 0-1)}

\item{FATOSK}{radiation configuration factor to sky (fractional, 0-1)}

\item{FATOBJ}{radiation configuration factor to ground (fractional, 0-1)}

\item{ZEN}{solar zenith angle (radians)}

\item{PDIF}{proportion of solar radiation that is diffuse (fractional, 0-1)}

\item{SHADE}{shade (\%)}

\item{postur}{postural orientation to sun, 1 = perpendicular, 2 = parallel, 0 = half way between}
}
\description{
R version of Fortran SOLAR.f (ectotherm model) for calculating absorbed solar radiation.
}
