% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR.R
\name{SOLAR_ENDO}
\alias{SOLAR_ENDO}
\title{SOLAR_ENDO}
\usage{
SOLAR_ENDO(AREATOTL, ABSAND, ABSANV, ABSSB, ASILN, PCTDIF, QNORM, SHADE,
  QSOLR, FASKY, FAVEG)
}
\arguments{
\item{AREATOTL}{A}

\item{ABSAND}{A}

\item{ABSANV}{A}

\item{ABSSB}{A}

\item{ASILN}{A}

\item{PCTDIF}{A}

\item{QNORM}{A}

\item{SHADE}{A}

\item{QSOLR}{A}

\item{FASKY}{A}

\item{FAVEG}{A}

\item{FATOBJ}{A}
}
\description{
R wrapper for Fortran binary of SOLAR_ENDO (endotherm model)
}
