% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLVENDO.R
\encoding{UTF-8}
\name{SOLVENDO}
\alias{SOLVENDO}
\title{SOLVENDO}
\usage{
SOLVENDO(SOLVENDO.input)
}
\arguments{
\item{SOLVENDO.input}{input vector}
}
\description{
R wrapper for Fortran binary of SOLVENDO (endotherm model)
}
