% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_behav.R
\encoding{UTF-8}
\name{trans_behav}
\alias{trans_behav}
\title{trans_behav - thermoregulatory model using a transient heat budget}
\usage{
trans_behav(Ww_g = 500, NGEOM = 4, GMREF = 3, FURTHRMK = 0, ZFURD = 2E-03, ZFURV = 2E-03, TC = 37, TCMAX = 45, TA = 20, TGRD = TA, TSKY = TA, VEL = 0.1, RH = 5, QSOLR = 0, Z = 20, SHADE = 0, NITESHAD = 0,...)
}
\arguments{
\item{t}{time intervals (s) at which output is required}

\item{Tc_init}{= Tairf(1), initial temperature (deg C) Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see details)}

\item{Ts_init}{= Tc_init + 0.1, initial shell temperature (deg C)}

\item{To_init}{= Tc_init + 0.2, initial surface temperature (deg C)}

\item{Ww_g}{= 500, weight (g)}

\item{T_F_min}{= 33, # user-specified fur thermal conductivity (W/mK), not used if 0}

\item{T_F_max}{= 38, # fur depth, dorsal (m)}

\item{T_B_min}{= 25, # fur depth, ventral (m)}

\item{CT_max}{= 43, # core temperature (°C)}

\item{rho_body}{animal density (kg/m3)}

\item{x_shell}{= 0.001, shell thickness, m}

\item{lump}{= 1, one lump (1) or two lump (2) model?}

\item{q}{metabolic rate (W/m3)}

\item{c_body}{= 3073, specific heat of flesh (J/kg-C)}

\item{c_body_inner}{= c_body, Specific heat of flesh J/(kg-K)}

\item{c_body_outer}{= c_body, Specific heat of outer shell J/(kg-K)}

\item{k_flesh}{= 0.5, conductivity of flesh (W/mK)}

\item{k_inner}{= k_flesh, Thermal conductivity of inner shell (W/mK, range: 0.412-2.8)}

\item{k_outer}{= k_flesh, Thermal conductivity of outer shell (W/mK, range: 0.412-2.8)}

\item{emis}{= 0.95, emissivity of skin (-)}

\item{alpha}{= 0.85, animal solar absorptivity (-)}

\item{geom}{= 2, Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see parameter 'shape_coeffs')}

\item{shape_b}{= 1/5, Proportionality factor (-) for going from volume to area, represents ratio of width:height for a plate, length:diameter for cylinder, b axis:a axis for ellipsoid}

\item{shape_c}{= 1/5, Proportionality factor (-) for going from volume to area, represents ratio of length:height for a plate, c axis:a axis for ellipsoid}

\item{shape_coefs}{= c(10.4713,.688,0.425,0.85,3.798,.683,0.694,.743), Custom surface area coefficients. Operates if posture = 5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*Ww_g^b, where AREA is in cm2 and Ww_g is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}

\item{posture}{= 'n', pointing normal 'n' or parallel 'p' to the sun's rays, or average 'b'?}

\item{orient}{= 1, does the object orient toward the sun? (0,1)}

\item{fatosk}{= 0.4, solar configuration factor to sky (-)}

\item{fatosb}{= 0.4, solar configuration factor to substrate (-)}

\item{alpha_sub}{= 0.2, substrate solar reflectivity, decimal percent}

\item{pctdif}{= 0.1, proportion of solar energy that is diffuse (rather than direct beam)}

\item{Tairf_sun}{air temperature function with time in sun, generated by 'approxfun' (deg C)}

\item{Tradf_sun}{radiant temperature function with time in sun, generated by 'approxfun'(deg C), averaging ground and sky}

\item{Tradf_shd}{radiant temperature function with time in shade, generated by 'approxfun'(deg C), averaging ground and sky}

\item{velf_sun}{wind speed function with time in sun, generated by 'approxfun' (m/s)}

\item{velf_shd}{wind speed function with time in shade, generated by 'approxfun' (m/s)}

\item{Qsolf_sun}{solar radiation function with time in sun, generated by 'approxfun' (W/m2)}

\item{Qsolf_shd}{solar radiation function with time in shade, generated by 'approxfun' (W/m2)}

\item{Zenf}{zenith angle of sun function with time, generated by 'approxfun' (90 is below horizon), degrees}

\item{press}{air pressure (Pa)}

\item{Tairf_shade}{air temperature function with time in shade, generated by 'approxfun' (deg C)}
}
\value{
Tc Core temperature (deg C)

Ts 'Skin' / shell temperature (deg C)

To Outer temperature (deg C)

Tcf Final (steady state) temperature (deg C), if conditions remained constant indefinately
}
\description{
This model uses the transient heat budget models (i.e. accounting for heat storage and
hence lag-effects of body size) to simulate thermoregulatory behaviour. .
}
