% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_metabolism.R
\name{water_metabolism}
\alias{water_metabolism}
\title{Estimate Metabolic Energy and Water Flux from Diet Composition}
\usage{
water_metabolism(
  p_prot = 0.44,
  p_fat = 0.48,
  p_carb = 0.08 * 0.8,
  p_dry = 0.3,
  kap_X = 0.8,
  mrate = mrate,
  p_X = mrate * 2,
  p_urea = 0.1,
  p_fecalH2O = 0.8
)
}
\arguments{
\item{p_prot}{Proportion of protein in dry food (default = 0.44).}

\item{p_fat}{Proportion of fat in dry food (default = 0.48).}

\item{p_carb}{Proportion of carbohydrate in dry food (default = 0.08 * 0.8).}

\item{p_dry}{Proportion of dry matter in wet food (default = 0.3).}

\item{kap_X}{Digestive efficiency (default = 0.8).}

\item{mrate}{Metabolic rate in J/time.}

\item{p_X}{Feeding rate in J/time (default = `mrate * 2`).}

\item{p_urea}{Proportion of urea in nitrogenous waste (default = 0.1).}

\item{p_fecalH2O}{Proportion of water in faeces (default = 0.8).}
}
\value{
A list containing:
\describe{
  \item{Q_avail}{Net energy available after metabolism (J/time).}
  \item{H2O_avail}{Net water available after accounting for excretion (g/time).}
  \item{H2OMet_g}{Metabolic water generated (g/time).}
  \item{H2OFaeces_g}{Water lost via faeces (g/time).}
  \item{WetFaeces_G}{Total wet mass of faeces (g/time).}
  \item{H2OUrine_g}{Water lost via urine (g/time).}
  \item{H2OFree_g}{Free water absorbed from wet food (g/time).}
  \item{WetFood_g}{Total wet food consumed (g/time).}
  \item{DryFood_g}{Total dry food consumed (g/time).}
}
}
\description{
Calculates energy availability and water gain/loss from digestion and excretion
based on dietary composition and metabolic rate.
}
\details{
Macronutrient metabolism produces metabolic water and contributes to energy
gain. This function estimates energy available after subtracting metabolic
requirements and calculates water gained from food and lost through excretion.
}
\examples{
water_metabolism(mrate = 5000)

}
