% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATERPROP.R
\encoding{UTF-8}
\name{WATERPROP}
\alias{WATERPROP}
\title{WATERPROP}
\usage{
WATERPROP(TWATR)
}
\arguments{
\item{TWATR}{water temperature (°C)}
}
\description{
R version of Fortran WATERPROP.f (ectotherm model) for calculating thermal properties of water.
}
