% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weathers1976Fig1.R
\docType{data}
\name{Weathers1976Fig1}
\alias{Weathers1976Fig1}
\title{Figure 1 from Weathers & Schoenbaechler 1976.}
\format{
A data frame with 47 rows and 3 columns:
\describe{
  \item{Time}{time of day that measurement occurred, i.e. day or night}
  \item{Tair}{chamber air temperature, °C}
  \item{mlO2gh}{ml O2 consumed per gram per hour}
}
}
\source{
Weathers, W. W., and D. C. Schoenbaechler. 1976. Regulation of body temperature in the Budgerygah, Melopsittacus undulatus. Australian Journal of Zoology 24:39–47.
}
\usage{
Weathers1976Fig1
}
\description{
Relation of oxygen consumption to ambient temperature during October-November for postabsorptive budgerigars resting in the dark
}
\keyword{datasets}
