% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weathers1976Fig2.R
\docType{data}
\name{Weathers1976Fig2}
\alias{Weathers1976Fig2}
\title{Figure 2b from Weathers & Schoenbaechler 1976.}
\format{
A data frame with 52 rows and 3 columns:
\describe{
  \item{Time}{time of day that measurement occurred, i.e. day or night}
  \item{Tair}{chamber air temperature, °C}
  \item{Tb}{body temperature, °C}
}
}
\source{
Weathers, W. W., and D. C. Schoenbaechler. 1976. Regulation of body temperature in the Budgerygah, Melopsittacus undulatus. Australian Journal of Zoology 24:39–47.
}
\usage{
Weathers1976Fig2
}
\description{
Relation of body temperature to ambient temperature for the budgerigars
}
\keyword{datasets}
