% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weathers1976Fig5.R
\docType{data}
\name{Weathers1976Fig5}
\alias{Weathers1976Fig5}
\title{Figure 5 from Weathers & Schoenbaechler 1976.}
\format{A data frame with 31 rows and 2 columns:
\describe{
  \item{Tair}{chamber air temperature, °C}
  \item{breaths_min}{breaths per minute}
}}
\source{
Weathers, W. W., and D. C. Schoenbaechler. 1976. Regulation of body temperature in the Budgerygah, Melopsittacus undulatus. Australian Journal of Zoology 24:39–47.
}
\usage{
Weathers1976Fig5
}
\description{
Relation of respiratory rate to ambient temperature in two budgerigars
}
\keyword{datasets}
