% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WETAIR.dp.R
\name{WETAIR.dp}
\alias{WETAIR.dp}
\title{WETAIR.dp}
\usage{
WETAIR.dp(db = db, dp = dp, bp = 101325)
}
\arguments{
\item{db}{Dry bulb temperature (degrees C)}

\item{dp}{Dew point temperature (degrees C)}

\item{bp}{Barometric pressure (pascal)}
}
\value{
e Vapour pressure (Pa)

esat Saturation vapour pressure (Pa)

vd Vapour density (kg m-3)

rw Mixing ratio (kg kg-1)

tvir Virtual temperature (K)

tvinc Virtual temperature increment (K)

denair Hourly predictions of the soil moisture under the maximum specified shade

cp Specific heat of air at constant pressure (J kg-1 K-1)

wtrpot Water potential (Pa)

Relative humidity (\%)
}
\description{
Calculates several properties of humid air as output variables below. The program
is based on equations from List, R. J. 1971. Smithsonian Meteorological Tables. Smithsonian
Institution Press. Washington, DC. WETAIR must be used in conjunction with function VAPPRS.
}
\details{
Input variables are shown below. The user must supply known values for dp, db and pb (bp at one standard
atmosphere is 101 325 pascals).
}
