% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBRAC_SHELTER.R
\name{ZBRAC_SHELTER}
\alias{ZBRAC_SHELTER}
\title{ZBRAC_SHELTER}
\usage{
ZBRAC_SHELTER(X1, X2, ZBRENT.in, results)
}
\arguments{
\item{X1}{A}

\item{X2}{A}

\item{ZBRENT.in}{A}

\item{TOL}{A}
}
\description{
R wrapper for Fortran binary of ZBRAC_SHELTER (endotherm model)
}
