% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBRENT.R
\name{ZBRENT}
\alias{ZBRENT}
\title{ZBRENT}
\usage{
ZBRENT(QM1, QM2, TOL, ZBRENT.in, results)
}
\arguments{
\item{QM1}{A}

\item{QM2}{A}

\item{TOL}{A}

\item{ZBRENT.in}{A}
}
\description{
R wrapper for Fortran binary of ZBRENT (endotherm model)
}
