% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBRENT.R
\name{ZBRENT_ENDO}
\alias{ZBRENT_ENDO}
\title{ZBRENT_ENDO}
\usage{
ZBRENT_ENDO(QM1, QM2, TOL, ZBRENT.in, results)
}
\arguments{
\item{QM1}{A}

\item{QM2}{A}

\item{TOL}{A}

\item{ZBRENT.in}{A}
}
\description{
R wrapper for Fortran binary of ZBRENT_ENDO (endotherm model)
}
