% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBRENT_SHELTER.R
\name{ZBRENT_SHELTER}
\alias{ZBRENT_SHELTER}
\title{ZBRENT_SHELTER}
\usage{
ZBRENT_SHELTER(X1, X2, TOL, ZBRENT.in)
}
\arguments{
\item{X1}{A}

\item{X2}{A}

\item{TOL}{A}

\item{ZBRENT.in}{A}
}
\description{
R wrapper for Fortran binary of ZBRENT_SHELTER (endotherm model)
}
