% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONV.R
\encoding{UTF-8}
\name{CONV_ENDO}
\alias{CONV_ENDO}
\title{CONV_ENDO}
\usage{
CONV_ENDO(TS, TENV, SHAPE, SURFAR, FLTYPE, FURTST, D, TFA, VEL, ZFUR, BP, ELEV)
}
\arguments{
\item{TS}{skin temperature (°C)}

\item{TENV}{fluid temperature at local height (°C)}

\item{SHAPE}{shape, 1 is cylinder, 2 is sphere, 3 is plate, 4 is ellipsoid}

\item{SURFAR}{surface area for convection, m2}

\item{FLTYPE}{FLUID TYPE: 0 = AIR; 1 = FRESH WATER; 2 = SALT WATER}

\item{FURTST}{test of fur presence (-) from IRPROP}

\item{D}{characteristic dimension for convection, m}

\item{TFA}{fur/air interface temperature (°C)}

\item{VEL}{wind speed (m/s)}

\item{ZFUR}{fur depth, mean (m) (from IRPROP)}

\item{BP}{barometric pressure (Pa), negative means altitude is used}

\item{ELEV}{elevation (m)}
}
\description{
R wrapper for Fortran binary of CONV_ENDO (endotherm model)
}
