% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ectotherm.R
\name{ectotherm}
\alias{ectotherm}
\title{Ectotherm model}
\usage{
ectotherm(amass, lometry, ABSMAX, ABSMIN, VTMAX, VTMIN, TBASK, TEMERGE, ctmax, ctmin,
 TPREF, dayact, nocturn, crepus, CkGrShad, burrow, climb, shdburrow, mindepth, maxdepth,
 M_1, M_2, M_3, skinwet, extref, DELTAR, ...)
}
\arguments{
\item{amass}{= 40, Mass of animal (g), note this model is 'steady state' so no lags in heating/cooling due to mass}

\item{lometry}{= 3, Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see details)}

\item{ABSMAX}{= 0.85, Maximum solar absorptivity, decimal percent}

\item{ABSMIN}{= 0.85, Maximum solar absorptivity, decimal percent}

\item{VTMAX}{= 34, Voluntary thermal maximum, degrees C (upper body temperature for foraging and also affects burrow depth selection)}

\item{VTMIN}{= 24, Voluntary thermal minimum, degrees C (lower body temperature for foraging)}

\item{TBASK}{= 17.5, Minimum basking temperature, degrees C}

\item{TEMERGE}{= 17.5, Temperature at which animal will move to a basking site, degrees C}

\item{TPREF}{= 30, Preferred body temperature, degrees C}

\item{ctmax}{= 40, Critical thermal maximum, degrees C (affects burrow depth selection)}

\item{ctmin}{= 6, Critical thermal minimum, degrees C (affects burrow depth selection)}

\item{dayact}{= 1, Diurnal activity allowed?  1=yes, 0=no}

\item{nocturn}{= 0, Nocturnal activity allowed?  1=yes, 0=no}

\item{crepus}{= 0, Crepuscular activity allowed?  1=yes, 0=no}

\item{CkGrShad}{= 1, shade seeking allowed?  1=yes, 0=no}

\item{burrow}{= 1 Shelter in burrow allowed?  1=yes, 0=no}

\item{climb}{= 0, climbing to seek cooler habitats allowed?  1=yes, 0=no}

\item{shdburrow}{= 0, choose if the animal's retreat is in the shade (1) or in the open (0)}

\item{mindepth}{= 2, Minimum depth (soil node #) to which animal can retreat if burrowing}

\item{maxdepth}{= 10, Maximum depth (soil node #) to which animal can retreat if burrowing}

\item{M_1}{= 0.013, Metabolic rate parameter 1 V_O2=M_1*M^M_2*10^(M_3*Tb) based on Eq. 2 from Andrews & Pough 1985. Physiol. Zool. 58:214-231}

\item{M_2}{= 0.800, Metabolic rate parameter 2}

\item{M_3}{= 0.038, Metabolic rate parameter 3}

\item{skinwet}{= 0.2, \% of surface area acting as a free-water exchanger, for computing cutaneous water loss}

\item{peyes}{= 0.03, \% of surface area taken up by open eye, for computing ocular water loss (only when active)}

\item{extref}{= 20, \% oxygen extraction efficiency, for respiratory water loss}

\item{DELTAR}{= 0.1, Temperature difference (deg C) between expired and inspired air, for computing respiratory water loss}
}
\description{
An implementation of the Niche Mapper ectotherm model that computes body temperature, water loss,
activity and microhabitat selection. It optionally runs the Dynamic Energy Budget (DEB) model for
computing mass budgets (inc. water budgets) and growth, development, reproduction trajectories
as constrained by food, activity and temperature (see Details). When not running the DEB model
a user-specified mass is used as well as a allometric (mass and body temperature) function to
compute metabolic rate. \cr\cr NOTE: The microclimate model, e.g. \code{\link{micro_global}}, must be run prior to running the ectotherm model
}
\details{
\strong{ Parameters controling how the model runs:}
\itemize{
\item{\code{microin}{ = "none", Directory where precomputed microclimate outputs (.csv files) are saved - if "none", assumes results are in list 'micro'}\cr}
\item{\code{nyears}{ = micro$nyears, Number of years the simulation runs for - must be consistent with dimensions of environmental input data}\cr}
\item{\code{ystrt}{ = 0, Year of the simulation to start at - will loop back over earlier periods at the end, useful for exploring cohort effects when running a life cycle with the DEB model}\cr}
\item{\code{enberr}{ = 0.0002, Factor by which the mass is multiplied to obtain a tolerance level for the heat budget solution}\cr}
\item{\code{live=}{ = 1, Live (metabolism/behaviour) or dead animal?}\cr}
\item{\code{write_input}{ = 0, Write csv files of final input to folder 'csv input' in working directory? 1 = yes, 0 = no}\cr}
\item{\code{write_csv}{ = 0, Write csv files of final output? 0 = no, 1= just yearout and yearsout, 2 = all output}\cr}
\item{\code{startday}{ = 1, Julian day of year at which simulation starts}\cr}
}
\strong{ Environmental inputs:}

\itemize{
\item{\code{minshade}{ = 0., Minimum shade (\%) available to the animal}\cr}
\item{\code{maxshades}{ = micro$MAXSHADES, Vector of daily maximum shade values}\cr}
\item{\code{FLTYPE}{ = 0.0, Fluid type 0=air, 1=water }\cr}
\item{\code{SUBTK}{ = 2.79, Substrate thermal conductivity (W/mC)}\cr}
\item{\code{REFL}{ = micro$REFL, Vector of daily substrate reflectances (\%)}\cr}
\item{\code{DEP}{ = micro$DEP,}\cr}
\item{\code{metout}{ = micro$metout, Microclimate model output for above ground, minimum shade conditions}\cr}
\item{\code{shadmet}{ = micro$shadmet, Microclimate model output for above ground, maximum shade conditions}\cr}
\item{\code{soil}{ = micro$soil, Microclimate model output for soil temperature, minimum shade conditions}\cr}
\item{\code{shadsoil}{ = micro$shadsoil, Microclimate model output for soil temperature, maximum shade conditions}\cr}
\item{\code{soilmoist}{ = micro$soilmoist, Microclimate model output for soil moisture, minimum shade conditions}\cr}
\item{\code{shadmoist}{ = micro$shadmoist, Microclimate model output for soil moisture, maximum shade conditions}\cr}
\item{\code{humid}{ = micro$humid, Microclimate model output for soil humidity, minimum shade conditions}\cr}
\item{\code{shadhumid}{ = micro$shadhumid, Microclimate model output for soil humidity, maximum shade conditions}\cr}
\item{\code{soilpot}{ = micro$soilpot, Microclimate model output for soil water potential, minimum shade conditions}\cr}
\item{\code{shadpot}{ = micro$shadpot, Microclimate model output for soil water potential, maximum shade conditions}\cr}
\item{\code{RAINFALL}{ = micro$RAINFALL, Vector of daily rainfall}\cr}
\item{\code{ectoin}{ = rbind(as.numeric(micro$ALTT),as.numeric(micro$REFL)[1],micro$longlat[1],micro$longlat[2]), Other items needed by the model - this needs to be tidied up}\cr}
}
\strong{ Morphological parameters:}

\itemize{
\item{\code{customallom}{ = c(10.4713,.688,0.425,0.85,3.798,.683,0.694,.743), Custom allometry coefficients. Operates if lometry=5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*mass^b, where AREA is in cm2 and mass is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}\cr}
\item{\code{shape_a}{ = 1., Proportionality factor (-) for going from volume to area, keep this 1 (redundant parameter that should be removed)}\cr}
\item{\code{shape_b}{ = 3, Proportionality factor (-) for going from volume to area, represents ratio of width:height for a plate, length:diameter for cylinder, b axis:a axis for ellipsoid }\cr}
\item{\code{shape_c}{ = 0.6666666667, Proportionality factor (-) for going from volume to area, represents ratio of length:height for a plate, c axis:a axis for ellipsoid}\cr}
\item{\code{FATOSK}{ = 0.4, Configuration factor to sky (-) for infrared calculations}\cr}
\item{\code{FATOSB}{ = 0.4, Configuration factor to subsrate for infrared calculations}\cr}
\item{\code{rinsul}{ = 0., Insulative fat layer thickness (m)}\cr}
\item{\code{ptcond}{ = 0.25, Fraction of surface contacting the substrate}\cr}
\item{\code{Spheat}{ = 3073, Specific heat of flesh J/(kg-K)}\cr}
\item{\code{Flshcond}{ = 0.5, Thermal conductivity of flesh (W/mC, range: 0.412-2.8)}\cr}
\item{\code{Andens}{ = 1000, Density of flesh (kg/m3)}\cr}
\item{\code{EMISAN}{ = 0.95, Emissivity of animal (0-1)}\cr}
}
\strong{ Behavioural parameters:}

\itemize{
\item{\code{warmsig}{ = 0, Warming signal for emergence? 1=yes, 0=no (if in burrow deeper than node 2, 0.1 degree sensitivity)}\cr}
\item{\code{fosorial}{ = 0, Fossorial activity? 1=yes, 0=no (this option hasn't been tested)}\cr}
\item{\code{rainact}{ = 0, Activity is limited by rainfall? 1=yes, 0=no, threshold rainfall for activity set by \code{actrainthresh}}\cr}
\item{\code{actrainthresh}{ = 0.1, Threshold (mm) of rain causing activity if \code{rainact}=1}\cr}
\item{\code{soilnode}{ = 4., Soil node (1-10, corresponding to values in \code{DEP}) at which eggs are laid (overridden if \code{frogbreed}=1)}\cr}
\item{\code{aquabask}{ = 0, If aquatic, does it bask? 0=no, stay at water temp, 1=yes, when not hungry, 2=all the time}\cr}
}
\strong{ Thermal physiological parameters:}

\itemize{
\item{\code{ctminthresh}{ = 12, Number of consecutive hours below CTmin that leads to death - simulation will terminate beyond this threshold if \code{ctkill}=1}\cr}
\item{\code{ctkill}{ = 0, Animal dies when it hits critical thermal limits? 1=yes, 0=no}\cr}
}
\strong{ Water and food budget parameters (only relevant if \code{DEB}=1):}

\itemize{
\item{\code{PFEWAT}{ = 73, Fecal water content (\%)}\cr}
\item{\code{PTUREA}{ = 0, Water in excreted nitrogenous waste (\%)}\cr}
\item{\code{FoodWater}{ = 82, Water content of food (\%)}\cr}
\item{\code{minwater}{ = 15, Minimum tolerated dehydration (\% of wet mass) - prohibits foraging if greater than this}\cr}
\item{\code{dessdeath}{ = 35, Maximum tolerated dehydration (\% of wet mass) - causes death if greater than this}\cr}
\item{\code{gutfill}{ = 75.,}\cr}
\item{\code{raindrink}{ = 0., Gut fill (\%) at which satiation occurs - if greater than 100\%, animal always tries to forage}\cr}
}
\strong{ Dynamic Energy Budget (DEB) model parameters:}
\itemize{
\item{\code{DEB}{ = 0, Run the DEB model (1) or just heat balance (0). Latter uses allometrically predicted respiration base on \code{M_1}, \code{M_2} and \code{M_3}}\cr}
\item{\code{fract}{ = 1, Scaling factor for DEB body-size covariation relationships - use it to make a metabolically scaled larger or smaller version of your animal}\cr}
}
\strong{ Core DEB parameters:}
\itemize{
\item{\code{z}{ = 2.825*fract, Zoom factor (cm)}\cr}
\item{\code{del_M}{ =  0.2144, Shape coefficient (-)}\cr}
\item{\code{F_m}{ = 12420, Surface area-specific maximum feeding rate J/cm2/h}\cr}
\item{\code{kap_X}{ = 0.85, Digestive efficiency (decimal \%)}\cr}
\item{\code{v}{ = 0.02795/24, Energy conductance (cm/h)}\cr}
\item{\code{kap}{ = 0.8206, fraction of mobilised reserve allocated to soma}\cr}
\item{\code{p_M}{ = 48.81/24, Volume-specific somatic maintenance (J/cm3/h)}\cr}
\item{\code{E_G}{ = 7512, Cost of structure (J/cm3)}\cr}
\item{\code{kap_R}{ = 0.95, Fraction of reproduction energy fixed in eggs}\cr}
\item{\code{k_J}{ = 0.006498/24, Maturity maintenance rate coefficient (1/h)}\cr}
\item{\code{E_Hb}{ = 866.6*fract^3, Maturity at birth (J)}\cr}
\item{\code{E_Hj}{ = E_Hb*fract^3, Maturity at metamorphosis (J)}\cr}
\item{\code{E_Hp}{ = 1.019e+04*fract^3, Maturity at puberty}\cr}
\item{\code{h_a}{ = 1.051e-08/(24^2), Weibull ageing acceleration (1/h2)}\cr}
\item{\code{s_G}{ = 0.01, Gompertz stress coefficient (-)}\cr}
\item{\code{E_0}{ = 9220*fract^4, Energy content of the egg (derived from core parameters) (J)}\cr}
}
\strong{ Thermal DEB parameters:}
\itemize{
\item{\code{T_REF}{ = 20, Reference temperature for rate correction (deg C)}\cr}
\item{\code{TA}{ = 8817 Arhhenius temperature}\cr}
\item{\code{TAL}{ = 50000, Arrhenius temperature for decrease below lower boundary of tolerance range \code{TL}}\cr}
\item{\code{TAH}{ = 90000, Arrhenius temperature for decrease above upper boundary of tolerance range \code{TH}}\cr}
\item{\code{TL}{ = 279, Lower boundary (K) of temperature tolerance range for Arrhenius thermal response}\cr}
\item{\code{TH}{ = 306, Upper boundary (K) of temperature tolerance range for Arrhenius thermal response}\cr}
}
\strong{ Compound/derived DEB parameters:}
\itemize{
\item{\code{E_m}{ = (p_M*z/kap)/v, Maximum reserve density (J/cm3)}\cr}
}
\strong{ Food-related axilliary DEB parameters:}
\itemize{
\item{\code{f}{ = 1, functional response (-), usually kept at 1 because gut model controls food availability such that f=0 when gut empty}\cr}
\item{\code{E_sm}{ = 350, Maximum volume-specific energy density of stomach (J/cm3)}\cr}
\item{\code{K}{ = 1, Half saturation constant (#/cm2)}\cr}
\item{\code{X}{ = 10, Food density (J/cm2)}\cr}
\item{\code{plantsim}{ = 0, parameters for plant growth simulation model - 0 is off, otherwise vector of 6 values: 1) shallowest soil node to use, 2) deepest soil node to use, 3) growth_delay (days), 4) water potential at wilting threshold (J/kg), 5) water potential at permanent wilting point (J/kg), 6) percent water in food at maximum}\cr}
}
\strong{ Composition-related axilliary DEB parameters:}
\itemize{
\item{\code{andens_deb}{ = Andens/1000, Animal density (g/cm3)}\cr}
\item{\code{d_V}{ = 0.3, Dry mass fraction of structure}\cr}
\item{\code{d_E}{ = 0.3, Dry mass fraction of reserve}\cr}
\item{\code{d_Egg}{ = 0.3, Dry mass fraction of egg}\cr}
\item{\code{mu_X}{ = 525000, Molar Gibbs energy (chemical potential) of food (J/mol)}\cr}
\item{\code{mu_E}{ = 585000, Molar Gibbs energy (chemical potential) of reserve (J/mol)}\cr}
\item{\code{mu_V}{ = 500000, Molar Gibbs energy (chemical potential) of structure (J/mol)}\cr}
\item{\code{mu_P}{ = 480000, Molar Gibbs energy (chemical potential) of faeces (J/mol)}\cr}
\item{\code{kap_X_P}{ = 0.1, Faecation efficiency of food to faeces (-)}\cr}
\item{\code{n_X}{ = c(1,1.8,0.5,0.15), chem. indices of C, O, H and N in food}\cr}
\item{\code{n_E}{ = c(1,1.8,0.5,0.15), chem. indices of C, O, H and N in reserve}\cr}
\item{\code{n_V}{ = c(1,1.8,0.5,0.15), chem. indices of C, O, H and N in structure}\cr}
\item{\code{n_P}{ = c(1,1.8,0.5,0.15), chem. indices of C, O, H and N in faeces}\cr}
\item{\code{n_M_nitro}{ = c(1,4/5,3/5,4/5), chem. indices of C, O, H and N in nitrogenous waste}\cr}
}
\strong{ Insect DEB model parameters (not yet in operation):}
\itemize{
\item{\code{metab_mode}{ = 0, Run insect model? 0 = no, 1 = hemimetabolus model (to do), 2 = holometabolous model}\cr}
\item{\code{stages}{ = 7, number of stages (max = 8) = number of instars plus 1 for egg + 1 for pupa + 1 for imago}\cr}
\item{\code{y_EV_l}{ = 0.95, yield of imago reserve on larval structure (mol/mol)}\cr}
\item{\code{S_instar}{ = c(2.660,2.310,1.916,0), stress at instar n: L_n^2/ L_n-1^2 (-)}\cr}
\item{\code{s_j}{ = 0.999, Reprod buffer/structure ratio at pupation as fraction of max}\cr}
}
\strong{ Inital conditions for DEB model:}
\itemize{
\item{\code{v_init}{ = 3e-9, Initial structural volume (cm3)}\cr}
\item{\code{E_init}{ = E_0/v_init, Initial reserve density (J/cm3)}\cr}
\item{\code{E_H_init}{ = 0, Initial maturity (J)}\cr}
\item{\code{stage}{ = 0, Initial stage (0=embryo, 1=juvenile, 2=mature but not yet reproducing, 3=beyond first reproduction)}\cr}
}
\strong{ Metabolic depression parameters (not yet functional):}
\itemize{
\item{\code{aestivate}{ = 0, Does the animal aestivate/go into torpor? 1=yes, 0=no}\cr}
\item{\code{depress}{ = 0.3, Fraction by which \code{p_M}, \code{k_J} and \code{v} are reduced during torpor}\cr}
}
\strong{ Reproductive phenology model parameters:}
\itemize{
\item{\code{clutchsize}{ = 5, Clutch size (#), overridden by \code{clutch_ab}}\cr}
\item{\code{clutch_ab}{ = c(0,0), # paramters for relationship between length (cm) and clutch size: clutch size = a*SVL-b, make a and b zero if fixed clutch size}\cr}
\item{\code{viviparous}{ = 0, Viviparous reproduction? 1=yes, 0=no (if yes, animal will be held in adult-sided female's body for duration of development and will experience her body temperature}\cr}
\item{\code{minclutch}{ = 0, Minimum clutch size if not enough in reproduction buffer for clutch size predicted by \code{clutch_ab} - if zero, will not operate}\cr}
\item{\code{batch}{ = 1, Invoke Pequerie et al.'s batch laying model?}\cr}
\item{\code{photostart}{ = 3, Photoperiod response triggering ovulation, none (0), summer solstice (1), autumnal equinox (2), winter solstice (3), vernal equinox (4), specified daylength thresholds (5 - uses \code{daylengthstart} and \code{daylengthfinish})}\cr}
\item{\code{photofinish}{ = 1, Photoperiod terminating ovulation, none (0), summer solstice (1), autumnal equinox (2), winter solstice (3), vernal equinox (4), specified daylength thresholds (5 - uses \code{daylengthstart} and \code{daylengthfinish})}\cr}
\item{\code{daylengthstart}{ = 12.5, Threshold daylength (h) for initiating breeding}\cr}
\item{\code{daylengthfinish}{ = 13., Threshold daylength (h) for terminating breeding}\cr}
\item{\code{photodirs }{ =  1, Is the start daylength trigger during a decrease (0) or increase (1) in day length?}\cr}
\item{\code{photodirf }{ =  0, Is the finish daylength trigger during a decrease (0) or increase (1) in day length?}\cr}
\item{\code{frogbreed}{ = 0, Amphibious animal breeding mode: 0 is off, 1 is exotrophic aquatic (eggs start when water present in container and within breeding season), 2 is exotrophic terrestrial/aquatic (eggs start at specified soil node within breeding season, diapause at birth threshold, start larval phase if water present in container), 3 endotrophic terrestrial (eggs start at specified soil node within breeding season and continue to metamorphosis on land), 4 turtle mode (eggs start at specified soil node within breeding season, hatch and animals enter water and stay there for the rest of their life, but leave the water if no water is present)}\cr}
\item{\code{frogstage}{ = 0, Life cycle control for amphibious animal: 0 is whole life cycle, 1 is just to metamorphosis (then reset and start again)}\cr}
\item{\code{reset}{ = 0, Life cycle reset options, 0=quit simulation upon death, 1=restart at emergence, 2=restart at first egg laid, 3=restart at end of breeding season, 4=reset at death}\cr}
\item{\code{breedactthresh}{ = 1, Threshold numbers of hours active after start of breeding season before eggs can be laid (simulating movement to the breeding site)}\cr}
\item{\code{breedrainthresh}{ = 0, Rain dependent breeder? 0 means no, otherwise enter rainfall threshold in mm}\cr}
\item{\code{breedtempthresh}{ = 200, Body temperature threshold below which breeding will occur}\cr}
\item{\code{breedtempcum}{ = 24*7, Cumulative time below temperature threshold for breeding \code{breedtempthresh} that will trigger breeding}\cr}
}
\strong{ Mortality rate parameters:}
\itemize{
\item{\code{ma}{ = 1e-4, Hourly active mortality rate (probability of mortality per hour)}\cr}
\item{\code{mi}{ = 0, Hourly inactive mortality rate (probability of mortality per hour)}\cr}
\item{\code{mh}{ = 0.5, Survivorship of hatchling in first year}\cr}
}
\strong{ Water body model parameters:}
\itemize{
\item{\code{container}{ = 0, Run the water body/container model? (aquatic start of life cycle, e.g. frog or mosquito)}\cr}
\item{\code{wetmod}{ = 0, Use precomputed wetland temperature \code{wetlandTemps} and depths \code{wetlandDepthds}?}\cr}
\item{\code{conth}{ = 10, Cylindrical container/pond height (cm)}\cr}
\item{\code{contw}{ = 100., Cylindrical container/pond diameter (cm)}\cr}
\item{\code{contype}{ = 1, Is 'containter' sitting on the surface, like a bucket (0) or sunk into the ground like a pond (1)}\cr}
\item{\code{rainmult}{ = 1, Rainfall multiplier to reflect catchment (don't make this zero unless you want a drought!)}\cr}
\item{\code{continit}{ = 0, Initial container water level (cm)}\cr}
\item{\code{conthole}{ = 0, Daily loss of height (mm) due to 'hole' in container (e.g. infiltration to soil, drawdown from water tank)}\cr}
\item{\code{contonly}{ = 1, Just run the container model and quit?}\cr}
\item{\code{contwet}{ = 80, \% of container surface acting as a free water exchanger}\cr}
\item{\code{wetlandTemps}{ = matrix(data = 0., nrow = 24*dim, ncol = 1), Matrix of hourly wetland temperaures (deg C)}\cr}
\item{\code{wetlandDepths}{ = matrix(data = 0., nrow = 24*dim, ncol = 1), Matrix of hourly wetland depths (cm)}\cr}
}
\strong{ Life stage-specific parameter allocation:}
\itemize{
\item{\code{thermal_stages}{ = matrix(data = c(rep(ctmin,8),rep(ctmax,8),rep(VTMIN,8),rep(VTMAX,8),rep(TBASK,8),rep(TPREF,8)), nrow = 8, ncol = 6), Stage specific thermal thresholds (ctmin,ctmax,VTMIN,VTMAX,TBASK,TPREF)}\cr}
\item{\code{behav_stages}{ = matrix(data = c(rep(dayact,8),rep(nocturn,8),rep(crepus,8),rep(burrow,8),rep(shdburrow,8),rep(mindepth,8),rep(maxdepth,8),rep(CkGrShad,8),rep(climb,8),rep(fosorial,8),rep(rainact,8),rep(actrainthresh,8),rep(breedactthresh,8),rep(flyer,8),rep(aquabask,8)), nrow = 8, ncol = 15), Stage specific behaviour dayact,nocturn,crepus,burrow,shdburrow,mindepth,maxdepth,CkGrShad,climb,fosorial,rainact,actrainthresh,breedactthresh,flyer,aquabask)}\cr}
\item{\code{water_stages}{ = matrix(data = c(rep(skinwet,8),rep(extref,8),rep(PFEWAT,8),rep(PTUREA,8),rep(FoodWater,8),rep(minwater,8),rep(raindrink,8),rep(gutfill,8)), nrow = 8, ncol = 8), Stage-specific water budget parameters (skinwet,extref,PFEWAT,PTUREA,FoodWater,minwater,raindrink,gutfill)}\cr}
\item{\code{arrhenius}{ = matrix(data = matrix(data = c(rep(TA,8),rep(TAL,8),rep(TAH,8),rep(TL,8),rep(TH,8)), nrow = 8, ncol = 5), nrow = 8, ncol = 5), Stage-specific 5-parameter Arrhenius thermal response for DEB model (TA,TAL,TAH,TL,TH)}\cr}
}
\strong{ Butterfly model parameters (not yet tested):}
\itemize{
\item{\code{wings}{ = 0, Turn wing model on? 1=yes, 0=no}\cr}
\item{\code{rho1_3}{ = 0.2, Wing reflectance (dec \%)}\cr}
\item{\code{trans1}{ = 0.00, Wing transmissivity (dec \%)}\cr}
\item{\code{aref}{ = 0.26, Width of surface #2 (cm) (back or horizontal or reference surface)}\cr}
\item{\code{bref}{ = 2.04, Common length (cm) where the two rectangles join}\cr}
\item{\code{cref}{ = 1.47, Width of surface #1 (cm) (wing)}\cr}
\item{\code{phi}{ = 179., Initial wing angle (degrees) (90 = vertical relative to body)}\cr}
\item{\code{phimax}{ =  phi, Maximum wing angle (degrees) (90 = vertical relative to body)}\cr}
\item{\code{phimin}{ =  phi, Minimum wing angle (degrees) (90 = vertical relative to body)}\cr}
\item{\code{flyer}{ = 0, Does the animal fly? 1=yes, 0=no}\cr}
\item{\code{flyspeed}{ = 5, Flying speed (m/s)}\cr}
\item{\code{flymetab}{ = 0.1035, Flight metabolic excess (W/g)}\cr}
}
\strong{Outputs:}

environ variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 YEAR - year of simulation
\item 3 DAY - day of simulation
\item 4 TIME - time of day (hours)
\item 5 TC - body temperature (deg C)
\item 6 SHADE - shade selected (\%)
\item 7 SOLAR  - solar radiation (W/m2) at animal location
\item 8 DEP - depth below ground (cm)
\item 9 ACT - activity state (0=inactive, 1=basking, 2=foraging)
\item 10 TA - air temperature (deg C) at animal location
\item 11 VEL - wind speed (m/s) at animal location
\item 12 RELHUM - relative humidity (\%) at animal location
\item 13 ZEN - zenith angle of sun (degrees - 90 = below the horizon)
\item 14 CONDEP - depth of water body (cm) (may not be simulated or supplied)
\item 15 WATERTEMP - temperature of water body (deg C) (may not be simulated or supplied)
\item 16 DAYLENGTH - day length (hours)
\item 17 WINGANGLE - wing angle (degrees) for butterfly model
\item 18 WINGTEMP - wing temperature (degrees C) for butterfly model
\item 19 FLYING - flying state (1=flying, 0=not flying) for butterfly model
\item 20 FLYTIME - flying time (hours) for butterfly model
}
enbal variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 YEAR - year of simulation
\item 3 DAY - day of simulation
\item 4 TIME - time of day (hours)
\item 5 QSOL - solar radiation absorbed (W)
\item 6 QIRIN - infrared radiation absorbed (W)
\item 7 QMET - metabolic heat production (W)
\item 8 QEVAP - evaporative heat loss (W)
\item 9 QIROUT - infrared radiation lost (W)
\item 10 QCONV - heat lost by convection (W)
\item 11 QCOND - heat lost by conduction (W)
\item 12 ENB - energy balance (degrees C)
\item 13 NTRY - iterations required for solution to heat balance equation
}
masbal variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 YEAR - year of simulation
\item 3 DAY - day of simulation
\item 4 TIME - time of day (hours)
\item 5 O2_ml - oxygen consumption rate (ml/h)
\item 6 CO2_ml - carbon dioxide production rate (ml/h)
\item 7 NWASTE_g - nitrogenous waste production (g/h)
\item 8 H2OFree_g - water from food (g/h)
\item 9 H2OMet_g - metabolic water production (g/h)
\item 10 DryFood_g - dry food intake (g/h)
\item 11 WetFood_g - wet foood intake (g/h)
\item 12 DryFaeces_g - dry faeces production (g/h)
\item 13 WetFaeces_G - wet faeces production (g/h)
\item 14 Urine_g - urine production (g/h)
\item 15 H2OResp_g - respiratory water loss (g/h)
\item 16 H2OCut_g - cutaneous water loss (g/h)
\item 17 H2OEye_g - ocular water loss (g/h)
\item 18 H2OBal_g - instantaneous water balance (g/h)
\item 19 H2OCumBal_g - cumulative water balance (g)
}
debout variables:
\itemize{
\item 1 JULDAY - day of year
\item 2 YEAR - year of simulation
\item 3 DAY - day of simulation
\item 4 TIME - time of day (hours)
\item 5 WETMASS - wet mass (g)
\item 6 E - reserve density (J/cm3)
\item 7 CUMREPRO - energy in reproduction buffer (J)
\item 8 HS - hazard rate (1/h)
\item 9 MASS_GUT - wet gut contents (g)
\item 10 SVL - length (mm) (might be Snout Vent Length but depends on choice of length measure for DEB paramter fitting)
\item 11 V - structural volume (cm3)
\item 12 E_H - maturity state (J)
\item 13 CUMBATCH - energy in batch for egg production (J)
\item 14 V_baby - structure of baby (cm3) (only if viviparous and pregnant
\item 15 E_baby - reserve density of baby (J/cm3) (only if viviparous and pregnant)
\item 16 Pregnant - pregnant? (only if viviparous) (0 or 1)
\item 17 Stage - life cycle stage (0=embryo, 1=juvenile, 2=mature but not yet reproducing, 3=beyond first reproduction)
\item 18 WETMASS_STD - wet mass (g) excluding stomach contents
\item 19 Body_cond - \% desiccated
\item 20 Surviv_Prob - survival probability due to joint influence of ageing and mortality rates
\item 21 Breeding - breeding state (1=breeding, 0=not breeding)
}
yearout variables:
\itemize{
\item 1 DEVTIME - development time (days)
\item 2 BIRTHDAY - birth day (day of year)
\item 3 BIRTHMASS - mass at birth (g)
\item 4 MONMATURE - months to maturity
\item 5 SVLREPRO - length (mm) at first reproduction
\item 6 FECUNDITY  - total fecundity
\item 7 CLUTCHES - total clutches
\item 8 MINRESERVE - minimum reserve density (J/cm3)
\item 9 LASTFOOD - food in last year (kg)
\item 10 TOTFOOD - total food eaten (kg)
\item 11 MINTB - minimum body temperature (deg C)
\item 12 MAXTB - maxmium body temperature (deg C)
\item 13 Pct_Dess - maximum level of desiccation (/%)
\item 14 LifeSpan - maximum life span (days)
\item 15 GenTime - generation time (years)
\item 16 R0 - net reproductive rate
\item 17 rmax - intrinsic rate of increase
\item 18 SVL - maximum length (mm)
}
yearsout variables:
\itemize{
\item 1 YEAR - year of simulation
\item 2 MaxStg - maximum stage reached in the year
\item 3 MaxWgt - maximum weight reached in the year (g)
\item 4 MaxLen - maximum length in the year (mm)
\item 5 Tmax - maximum annual body temperature (deg C)
\item 6 Tmin  - minimum annual body temperature (deg C)
\item 7 MinRes - minimum annual reserve density (J/cm3)
\item 8 MaxDes - maximum annual desiccation level (% of normal wet body mass)
\item 9 MinShade - minimum annual shade selected
\item 10 MaxShade - maximum annual shade selected
\item 11 MinDep - minimum annual depth selected (cm)
\item 12 MaxDep - maximum annual depth selected (cm)
\item 13 Bsk - annual basking hours
\item 14 Forage - annual foraging hours
\item 15 Dist - annual distance travelled (km)
\item 16 Food - annual food eaten (g)
\item 17 Drink - annual water drunk (g)
\item 18 NWaste - annual nitrogenous waste (g)
\item 19 O2 - annual O2 production (ml)
\item 20 Clutch - annual clutches (#)
\item 21 Fec - annual fecundity (#)
\item 22 CauseDeath - cause of death, 0=no death, 1=cold, 2=heat, 3=desiccation, 4=starvation, 5=ageing
\item 23 tLay - day of year at which eggs laid
\item 24 tEgg - day of year entering egg stage
\item 25 tStg1 - day of year entering life cycle stage 1
\item 26 tStg2 - day of year entering life cycle stage 2
\item 27 tStg3 - day of year entering life cycle stage 3
\item 28 tStg4 - day of year entering life cycle stage 4
\item 29 tStg5 - day of year entering life cycle stage 5
\item 30 tStg6 - day of year entering life cycle stage 6
\item 31 tStg7 - day of year entering life cycle stage 7
\item 32 tStg8 - day of year entering life cycle stage 8
\item 33 mStg1 - mass entering life cycle stage 1 (g)
\item 34 mStg2 - mass entering life cycle stage 2 (g)
\item 35 mStg3 - mass entering life cycle stage 3 (g)
\item 36 mStg4 - mass entering life cycle stage 4 (g)
\item 37 mStg5 - mass entering life cycle stage 5 (g)
\item 38 mStg6 - mass entering life cycle stage 6 (g)
\item 39 mStg7 - mass entering life cycle stage 7 (g)
\item 40 mStg8 - mass entering life cycle stage 8 (g)
\item 41 surviv - survival probability at end of given year
\item 42 ovip_surviv - breeding state (1=breeding, 0=not breeding)
\item 43 fitness - fecundity by survival probability at end of given year
\item 44 deathstage - life stage at which death occurred
}
}
\examples{
# run the microclimate model
micro<-micro_global(loc="Kuranda, Queensland")

# retrieve output
metout<-as.data.frame(micro$metout) # above ground microclimatic conditions, min shade
shadmet<-as.data.frame(micro$shadmet) # above ground microclimatic conditions, max shade
soil<-as.data.frame(micro$soil) # soil temperatures, minimum shade
shadsoil<-as.data.frame(micro$shadsoil) # soil temperatures, maximum shade

# append dates
days<-rep(seq(1,12),24)
days<-days[order(days)]
dates<-days+metout$TIME/60/24-1 # dates for hourly output
dates2<-seq(1,12,1) # dates for daily output
metout<-cbind(dates,metout)
soil<-cbind(dates,soil)
shadmet<-cbind(dates,shadmet)
shadsoil<-cbind(dates,shadsoil)

# run the ectotherm model
ecto<-ectotherm(VTMAX=35,VTMIN=30,TPREF=33,TBASK=20,TEMERGE=10)

# retrieve output
environ<-as.data.frame(ecto$environ) # activity, Tb and environment
enbal<-as.data.frame(ecto$enbal) # energy balance values
masbal<-as.data.frame(ecto$masbal) # mass balance value (note most missing if DEB model not running)

# append dates
environ<-cbind(dates,environ)
masbal<-cbind(dates,masbal)
enbal<-cbind(dates,enbal)

############### plot results ######################

# Hourly Tb (black), activity (orange, 5=bask, 10=forage), depth (brown, m) and shade (green, \%/10)
with(environ, plot(TC~dates,ylab="Tb, depth, activity and shade", xlab="month of year",
ylim=c(-20,70),type = "l"))
with(environ, points(ACT*5~dates,type = "l",col="orange"))
with(environ, points(SHADE/10~dates,type = "l",col="green"))
with(environ, points(DEP/10~dates,type = "l",col="brown"))
#with(metout, points(TAREF~dates,type = "l",col="light blue"))
abline(ecto$VTMAX,0,lty=2,col='red')
abline(ecto$VTMIN,0,lty=2,col='blue')

# seasonal activity plot (dark blue = night, light blue = basking, orange = foraging)
forage<-subset(environ,ACT==2)
bask<-subset(environ,ACT==1)
night<-subset(metout,ZEN==90)
day<-subset(metout,ZEN!=90)
with(night,plot(TIME/60~JULDAY,ylab="Hour of Day",xlab="Day of Year",pch=15,cex=2,col='dark blue'))
# nighttime hours
with(forage,points((TIME-1)~JULDAY,pch=15,cex=2,col='orange')) # foraging Tbs
with(bask,points((TIME-1)~JULDAY,pch=15,cex=2,col='light blue')) # basking Tbs
}

