% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_endo.R
\name{ellipsoid}
\alias{ellipsoid}
\title{Implementation of the endotherm model .}
\usage{
ellipsoid(posture = 4.5, mass = 0.5, coreT = 37, furdepth = 5,
  O2eff = 0.2, furcond = 0.04, airT = 20, windspd = 1, rh = 50)
}
\arguments{
\item{posture}{= 4.5, Shape, ratio of long to short axis of a prolate ellipsoid}

\item{mass}{= 0.5, Body Mass (kg)}

\item{coreT}{= 37, Core temperature (deg C)}

\item{furdepth}{= 5, Fur depth (mm)}

\item{O2eff}{= 0.2, Oxygen extraction efficiency (decimal \%)}

\item{furcond}{= 0.04, Conductivity of fur (W/Cm)}

\item{airT}{= 20, Air temperature (deg C)}

\item{windspd}{= 1, Wind speed (m/s)}

\item{rh}{= 50, Relative humidity (\%)}
}
\value{
AirTemp Air temperature for calculation (deg C)

Windspeed Wind speed for calculation (m/s)

RelHum Relative humidity for calculation (\%)

Tcore Core temperature for calculation (deg C)

UCT Upper Critical Temperature (deg C)

LCT Lower Critical Temperature (deg C)

Qresp_gph Respiration rate (W)

Qresp_W Respiration rate (W)

Qresp_kjph Respiration rate (kJ/h)

Tskin Skin temperature (deg C)

Qgen required heat generation (W)

QgenFinal Required heat generation capped at basal (W)

mlO2ph Oxygen consumption rate (ml/h)

PctBasal Metabolic rate (\% of basal)

status 1=cold, 2=happy, 3=hot

H2Oloss_W Watts of heat to be dumped

H2O_gph Water loss rate required to dump heat based on latent heat of vaporization

massph_percent Percent of body mass lost as water per hour

timetodeath Time to death (hours) from desiccation (15\% desiccated) if no water to drink
}
\description{
An implementation of described in
Porter, W. P., and M. Kearney. 2009. Size, shape, and the thermal niche of endotherms.
Proceedings of the National Academy of Sciences 106:19666-19672.
with additional rough calculations for water loss. Note this model is only relevant for
conditions where there is no solar radiation and air/ground/sky temperatures are equal
}
\examples{
micro<-micro_global(loc = 'Birdsville, Australia') # run the model with default location and settings

metout<-as.data.frame(micro$metout) # above ground microclimatic conditions, min shade
shadmet<-as.data.frame(micro$shadmet) # above ground microclimatic conditions, max shade
soil<-as.data.frame(micro$soil) # soil temperatures, minimum shade
shadsoil<-as.data.frame(micro$shadsoil) # soil temperatures, maximum shade

# append dates
days<-rep(seq(1,12),24)
days<-days[order(days)]
dates<-days+metout$TIME/60/24-1 # dates for hourly output
dates2<-seq(1,12,1) # dates for daily output
metout<-cbind(dates,metout)
soil<-cbind(dates,soil)
shadmet<-cbind(dates,shadmet)
shadsoil<-cbind(dates,shadsoil)

endo<-cbind(metout[,1:3],ellipsoid(airT = shadmet$TALOC, windspd = shadmet$VLOC, rh = shadmet$RHLOC))

with(endo,{plot(H2O_gph ~ dates,xlab = "Date and Time", ylab = "Water Loss Rate (g/h)"
, type = "l",main=paste("Evaporative Water Loss",sep=""))})
with(endo,{plot(QgenFinal ~ dates,xlab = "Date and Time", ylab = "Metabolic Rate (W)"
, type = "l",main=paste("Metabolic Heat Generation",sep=""))})
with(endo,{plot(PctBasal ~ dates,xlab = "Date and Time", ylab = "Metabolic Rate (\% of basal)"
, type = "l",main=paste("Metabolic Heat Generation",sep=""))})
with(endo,{plot(Tskin ~ dates,xlab = "Date and Time", ylab = "Skin and Core Temperature (deg C)"
, type = "l",main=paste("Skin and Core Temperature",sep=""),ylim=c(min(cbind(endo$Tcore,endo$Tskin)),
 max(cbind(endo$Tcore,endo$Tskin))))})
with(endo,{points(Tcore ~ dates,xlab = "Date and Time",lty=2, type = "l")})
with(endo,{plot(timetodeath ~ dates,xlab = "Date and Time", ylab = "Time to Death (h)"
, type = "l",main=paste("Time to Death by Desiccation",sep=""), ylim=c(0,24))})
}

