% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_ecto.R
\encoding{UTF-8}
\name{FUN_ecto}
\alias{FUN_ecto}
\title{FUN_ecto}
\usage{
FUN_ecto(
  AMASS = AMASS,
  GEOMETRY = GEOMETRY,
  ATOT = AREA,
  AV = AV,
  AT = AT,
  AL = AL,
  VOL = VOL,
  R = R,
  R1 = R1,
  RINSUL = RINSUL,
  ASEMAJR = ASEMAJR,
  BSEMINR = BSEMINR,
  CSEMINR = CSEMINR,
  M_1 = M_1,
  M_2 = M_2,
  M_3 = M_3,
  EXTREF = EXTREF,
  PANT = PANT,
  RQ = RQ,
  FLSHCOND = FLSHCOND,
  PSI_BODY = PSI_BODY,
  SKINW = SKINW,
  AEFF = AEFF,
  PEYES = PEYES,
  FATOSK = FATOSK,
  FATOSB = FATOSB,
  FATOBJ = FATOBJ,
  EMISAN = EMISAN,
  EMISSB = EMISSB,
  EMISSK = EMISSK,
  FLTYPE = FLTYPE,
  TA = TA,
  TSKY = TSKY,
  TSUBST = TSUBST,
  TGRD = TGRD,
  VEL = VEL,
  QSOLAR = QSOLAR,
  QIRIN = QIRIN,
  RELHUM = RELHUM,
  BP = BP,
  ALT = ALT,
  SUBTK = SUBTK,
  O2GAS = O2GAS,
  CO2GAS = CO2GAS,
  N2GAS = N2GAS,
  X
)
}
\arguments{
\item{AMASS}{body mass, kg}

\item{GEOMETRY}{organism shape, 0-5, 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom}

\item{ATOT}{total body surface area (m2)}

\item{AV}{ventral surface area (m2)}

\item{AT}{body surface area contacting another organism of same temperature (m2)}

\item{AL}{animal characteristic dimension (length) (m)}

\item{VOL}{body volume (m3)}

\item{R}{total body radius (m)}

\item{R1}{flesh radius (i.e. radius to start of insulation if present) (m)}

\item{RINSUL}{depth of insulation (m) (not yet used)}

\item{ASEMAJR}{length of semi-major radius for ellipsoid (m)}

\item{BSEMINR}{length of semi-minor radius 1 for ellipsoid (m)}

\item{CSEMINR}{length of semi-minor radius 1 for ellipsoid (m)}

\item{M_1}{metabolic rate parameter 1 V_O2=M_1*M^M_2*10^(M_3*Tb), in ml O2 / h, default parameters for lizards based on Eq. 2 from Andrews & Pough 1985. Physiol. Zool. 58:214-231}

\item{M_2}{metabolic rate parameter 2}

\item{M_3}{metabolic rate parameter 3}

\item{EXTREF}{oxygen extraction efficiency (\%)}

\item{PANT}{multiplier on breathing rate, for respiratory water loss via panting}

\item{RQ}{respiratory quotient (fractional, 0-1)}

\item{FLSHCOND}{thermal conductivity of flesh (W/mK)}

\item{PSI_BODY}{water potential of body (J/kg) - affects skin humidity for water vapour exchange}

\item{SKINW}{fraction of total surface area acting as a free water surface (fractional, 0-1)}

\item{AEFF}{effective area acting as a free-water exchanger, drives cutaneous evaporation (m2)}

\item{PEYES}{proportion of total surface area that is 'wet' eye (fractional, 0-1)}

\item{FATOSK}{configuration factor to sky (-)}

\item{FATOSB}{configuration factor to substrate (-)}

\item{FATOBJ}{configuration factor to nearby object (-) (not functional at the moment)}

\item{EMISAN}{emissivity of animal (fractional, 0-1)}

\item{EMISSB}{emissivity of substrate (fractional, 0-1)}

\item{EMISSK}{emissivity of sky (fractional, 0-1)}

\item{FLTYPE}{fluid type, air (0) or water (1)}

\item{TA}{air temperature (°C)}

\item{TSKY}{sky temperature (°C)}

\item{TSUBST}{substrate temperature (driving conduction) (°C)}

\item{TGRD}{ground temperature (driving radiation gain) (°C)}

\item{VEL}{wind speed (m/s)}

\item{QSOLAR}{solar radiation in, computed by SOLAR_ecto (W)}

\item{QIRIN}{longwave radiation in, computed by RADIN_ecto (W)}

\item{RELHUM}{relative humidity (\%)}

\item{BP}{air pressure (Pa)}

\item{ALT}{elevation (m)}

\item{SUBTK}{substrate thermal conductivity (W/mK)}

\item{O2GAS}{oxygen concentration of atmosphere (\%)}

\item{CO2GAS}{carbon dioxide concentration of atmosphere (\%)}

\item{X}{current guess of core body temperature (°C)}
}
\description{
R version of Fortran FUN.f (ectotherm model) for guessing a core body temperature that balances the heat budget.
}
