% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOM.R
\name{GEOM}
\alias{GEOM}
\title{GEOM}
\usage{
GEOM(AMASS, ANDENS, FATPCT, POSTUR, ZFUR, SUBQFAT, SHAPE_B, SHAPE_B_REF,
  SHAPE_C, DHARA, RHOARA, PTCOND, SAMODE, ORIENT)
}
\arguments{
\item{AMASS}{mass, kg}

\item{ANDENS}{volume, kg/m3}

\item{FATPCT}{fat percentage, \%}

\item{POSTUR}{shape, -}

\item{ZFUR}{fur depth, m}

\item{SUBQFAT}{subcutaneous fat presence, -}

\item{SHAPE_B}{multiplier on long axis for shape change, -}

\item{SHAPE_B_REF}{reference multiplier on long axis for shape change, -}

\item{SHAPE_C}{multiplier on width relative to height for plate shape change, -}

\item{DHARA}{hair diameter, m}

\item{RHOARA}{hair density, 1/m2}

\item{PTCOND}{percentage of surface area conducting to substrate, \%}

\item{SAMODE}{surface area mode, -}

\item{ORIENT}{orientation to solar, -}
}
\description{
R wrapper for Fortran binary of GEOM (endotherm model)
}
