% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_SHAPES.R
\encoding{UTF-8}
\name{GET_SHAPES}
\alias{GET_SHAPES}
\title{GET_SHAPES}
\usage{
GET_SHAPES(
  MASSs = c(5.0987, 33.567, 3.283, 10.854),
  AREA = 1.947321,
  SHAPE_Bs = c(1.58, 2.08, 7.69, 8),
  SHAPE_Bs.min = c(1.1, 2.3, 7.5, 6.5),
  SHAPE_Bs.max = c(1.7, 3.5, 10, 10),
  AREAFRACs = c(0.08291887, 0.3269846, 0.11025155, 0.18479669),
  DENSITYs = c(1050, 1050, 1050, 1050),
  FATPCTs = c(5, 36, 10, 23),
  SHAPEs = c(4, 1, 1, 1),
  SUBQFATs = c(1, 1, 1, 1),
  DHAIRDs = c(3e-05, 3e-05, 3e-05, 3e-05),
  DHAIRVs = c(3e-05, 3e-05, 3e-05, 3e-05),
  INSDENDs = c(3e+08, 3e+08, 3e+08, 3e+08),
  INSDENVs = c(3e+08, 3e+08, 3e+08, 3e+08),
  tol = 1e-06,
  maxiter = 10000
)
}
\arguments{
\item{MASSs}{mass per body part (head, trunk, arm, leg), kg}

\item{AREA}{body surface area, m2}

\item{SHAPE_Bs}{shape factors per body part, -}

\item{SHAPE_Bs.min}{min allowable SHAPE_Bs, -}

\item{SHAPE_Bs.max}{max allowable SHAPE_Bs, -}

\item{AREAFRACs}{area fraction per body part, -}

\item{DENSITYs}{density per body part, kg/m3}

\item{FATPCTs}{fat percentage per body part, /%}

\item{SHAPEs}{shape type per body part, -}

\item{SUBQFATs}{presence of subcutaneous fat per body part, -}

\item{DHAIRDs}{dorsal hair diameters, m}

\item{DHAIRVs}{ventral hair diameters, m}

\item{INSDENDs}{dorsal insulation depth, m}

\item{tol}{tolerance, -}

\item{maxiter}{maximum iterations,}

\item{SUBQFAT}{subcutaneous fat presence, -}

\item{INSDENvs}{ventral insulation depth, m}
}
\description{
R wrapper for Fortran binary of GET_SHAPES (HomoTherm model)
}
