% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPROP.R
\name{IRPROP}
\alias{IRPROP}
\title{IRPROP}
\usage{
IRPROP(TA, DHAIRD, DHAIRV, LHAIRD, LHAIRV, ZFURD, ZFURV, RHOD, RHOV, REFLD,
  REFLV, ZFURCOMP, PVEN, KHAIR)
}
\arguments{
\item{TA}{A}

\item{DHAIRD}{A}

\item{DHAIRV}{A}

\item{LHAIRD}{A}

\item{LHAIRV}{A}

\item{ZFURD}{A}

\item{ZFURV}{A}

\item{RHOD}{A}

\item{RHOV}{A}

\item{REFLD}{A}

\item{REFLV}{A}

\item{ZFURCOMP}{A}

\item{KHAIR}{A}

\item{SHAPE_B_MAX}{A}

\item{SHAP_B_REF}{A}

\item{SHAPE_B}{A}

\item{MAXPTVEN}{A}
}
\description{
R wrapper for Fortran binary of IRPROP (endotherm model)
}
