% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRPROP.R
\encoding{UTF-8}
\name{IRPROP}
\alias{IRPROP}
\title{IRPROP}
\usage{
IRPROP(
  TA,
  DHAIRD,
  DHAIRV,
  LHAIRD,
  LHAIRV,
  ZFURD,
  ZFURV,
  RHOD,
  RHOV,
  REFLD,
  REFLV,
  ZFURCOMP,
  PVEN,
  KHAIR
)
}
\arguments{
\item{TA}{air temperature at local height (°C)}

\item{DHAIRD}{hair diameter, dorsal (m)}

\item{DHAIRV}{hair diameter, ventral (m)}

\item{LHAIRD}{hair length, dorsal (m)}

\item{LHAIRV}{hair length, ventral (m)}

\item{ZFURD}{fur depth, dorsal (m)}

\item{ZFURV}{fur depth, ventral (m)}

\item{RHOD}{hair density, dorsal (1/m2)}

\item{RHOV}{hair density, ventral (1/m2)}

\item{REFLD}{fur reflectivity dorsal (fractional, 0-1)}

\item{REFLV}{fur reflectivity ventral (fractional, 0-1)}

\item{ZFURCOMP}{depth of compressed fur (for conduction) (m)}

\item{PVEN}{fraction of surface area that is ventral fur (fractional, 0-1)}

\item{KHAIR}{hair thermal conductivity (W/m°C)}
}
\description{
R wrapper for Fortran binary of IRPROP (endotherm model)
}
