% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_global.R
\name{micro_global}
\alias{micro_global}
\title{micro_global}
\usage{
micro_global(loc, timeinterval, nyears, ...)
}
\arguments{
\item{loc}{Either a longitude and latitude (decimal degrees) or a place name to search for on Google Earth}

\item{timeinterval}{The number of time intervals to generate predictions for over a year (must be 12 <= x <=365)}

\item{nyears}{The number of years to run}

\item{soiltype}{Soil type: Rock = 0, sand = 1, loamy sand = 2, sandy loam = 3, loam = 4, silt loam = 5, sandy clay loam = 6, clay loam = 7, silt clay loam = 8, sandy clay = 9, silty clay = 10, clay = 11, based on Campbell and Norman 1990 Table 9.1.}

\item{REFL}{Soil solar reflectance, decimal percent}

\item{slope}{Slope in degrees}

\item{aspect}{Aspect in degrees (0 = north)}

\item{DEP}{Soil depths at which calculations are to be made (cm), must be 10 values starting from 0, and more closely spaced near the surface}

\item{Usrhyt}{Local height at which air temperature, relative humidity and wind speed are to be computed (cm)}
}
\description{
An implementation of the Niche Mapper microcliate model that uses the global climate database
derived from "New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data
set of surface climate over global land areas. Climate Research 21:1-25"
It also optionally uses a global monthly soil moisture estimate from NOAA CPC Soil Moisture http://140.172.38.100/psd/thredds/catalog/Datasets/cpcsoil/catalog.html
Aerosol attenuation can also be computed based on the Global Aerosol Data Set (GADS)
Koepke, P., M. Hess, I. Schult, and E. P. Shettle. 1997. Global Aerosol Data Set. Max-Planck-Institut for Meteorologie, Hamburg
by choosing the option 'rungads<-1'
}

