% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onelump.R
\name{onelump}
\alias{onelump}
\title{One-lump Transient Heat Budget for Constant Environment}
\usage{
onelump(t, Tc_init, mass, geom, Tair, Trad, vel, Qsol, Zen, ...)
}
\arguments{
\item{t}{= seq(1,3600,60), time intervals (s) at which output is required}

\item{Tc_init}{= 5, initial temperature (deg C)}

\item{mass}{= 500, animal mass (g)}

\item{geom}{= 2, Organism shape, 0-5, Determines whether standard or custom shapes/surface area/volume relationships are used: 0=plate, 1=cyl, 2=ellips, 3=lizard (desert iguana), 4=frog (leopard frog), 5=custom (see parameter 'shape_coefs')}

\item{Tair}{= 30, air temperature (deg C)}

\item{Trad}{= 30, radiant temperature (deg C), averaging ground and sky}

\item{vel}{= 0.1, wind speed (m/s)}

\item{Qsol}{= 500, solar radiation (W/m2)}

\item{Zen}{= 20, zenith angle of sun (90 is below horizon), degrees}

\item{kflesh}{= 0.5, Thermal conductivity of flesh (W/mK, range: 0.412-2.8)}

\item{q}{= 0, metabolic heat production rate W/m3}

\item{cp}{= 3073, Specific heat of flesh J/(kg-K)}

\item{emis}{= 0.95, Emissivity of animal (0-1)}

\item{rho}{= 932, animal density (kg/m3)}

\item{abs}{= 0.85, solar absorptivity, decimal percent}

\item{shape_coefs}{= c(10.4713,.688,0.425,0.85,3.798,.683,0.694,.743), Custom shape coefficients. Operates if geom=5, and consists of 4 pairs of values representing the parameters a and b of a relationship AREA=a*mass^b, where AREA is in cm2 and mass is in g. The first pair are a and b for total surface area, then a and b for ventral area, then for sillhouette area normal to the sun, then sillhouette area perpendicular to the sun}

\item{shape_b}{= 1/5, Proportionality factor (-) for going from volume to area, represents ratio of width:height for a plate, length:diameter for cylinder, b axis:a axis for ellipsoid}

\item{shape_c}{= 1/5, Proportionality factor (-) for going from volume to area, represents ratio of length:height for a plate, c axis:a axis for ellipsoid}

\item{posture}{= 'n' pointing normal 'n', parallel 'p' to the sun's rays, or 'b' in between?}

\item{orient}{= 1, does the object orient toward the sun? (0,1)}

\item{fatosk}{= 0.4, Configuration factor to sky (-) for infrared calculations}

\item{fatosb}{= 0.4, Configuration factor to subsrate for infrared calculations}

\item{abs_sub}{= 0.2, substrate solar reflectivity, decimal percent}

\item{pctdif}{= 0.1, proportion of solar energy that is diffuse (rather than direct beam)}

\item{press}{= 101325, air pressure (Pa)}
}
\value{
Tc Core temperature (deg C)

Tcf Final (steady state) temperature (deg C), if conditions remained constant indefinately

tau Time constant (s)

dTc Rate of change of core temperature (deg C/s)
}
\description{
Transient, 'one-lump', heat budget for computing rate of change of temperature
under a constant environment
Michael Kearney, Raymond Huey and Warren Porter developed this R function and example in September 2017.
}
\examples{
library(NicheMapR)

# compare heating rates of a variety of different-sized objects

t=seq(1,3600*2,60) # times (in seconds) to report back values - every minute for 2 hrs
tmins <- t/60

par(mfrow = c(1,2))
mass <- 5
Tc_init <- 20
geom <- 2
Tair <- 20
Trad <- Tair
vel <- 1
Qsol <- 500
Zen <- 20
abs <- 0.85

Tbs<-onelump(t=t, abs = abs, Tc_init = Tc_init, mass = mass,
  geom = geom, Tair = Tair, Trad = Trad, vel = vel, Qsol = Qsol, Zen = Zen)
plot(Tbs$Tc ~ tmins, type= 'l' ,col = 1, ylim = c(20, 32), ylab = 'Temperature, deg C',xlab='time, s', las = 1)
text(80, 29, "    5000 g")
text(80, 25, "5 g")
text (40, 20.5, "Tair for both sizes", col = "blue")
text(90, 28, "   vel = 1.0 m/s")
text(90, 24, "     vel = 1.0 m/s")

mass <- 5000
Tbs<-onelump(t=t, abs = abs, Tc_init = Tc_init, mass = mass,
  geom = geom, Tair = Tair, Trad = Trad, vel = vel, Qsol = Qsol, Zen = Zen)
points(Tbs$Tc~tmins,type='l',lty = 2, col=1)
abline(Tair,0, lty = 1, col = 'blue')
abline(h = Tair + .1, lty = 2, col = 'blue')

mass <- 5
Tair <- 25
Trad <- Tair
vel <-0.5

Tbs<-onelump(t=t, abs = abs, Tc_init = Tc_init, mass = mass,
  geom = geom, Tair = Tair, Trad = Trad, vel = vel, Qsol = Qsol, Zen = Zen)
plot(Tbs$Tc~tmins,type='l',col=1,ylim=c(20,32),ylab='Temperature, deg C',xlab='time, s', las = 1)
abline(h = Tair, lty = 1, col = 'blue')

mass <- 5000
Tair <- 20
Trad <- Tair
vel <-1

Tbs<-onelump(t=t, abs = abs, Tc_init = Tc_init, mass = mass,
  geom = geom, Tair = Tair, Trad = Trad, vel = vel, Qsol = Qsol, Zen = Zen)
points(Tbs$Tc~tmins,type='l',lty = 2, col=1)
abline(h = Tair, lty = 2, col = 'blue')

text(65, 31, "5 g")
text(80, 29, "5000 g")
text(60, 20.5, "Tair for large animal", col = "blue")
text(63, 25.4, "Tair for small animal", col = "blue")
text(80, 30.1, "vel = 0.5 m/s")
text(93, 28, "vel = 1.0 m/s")
}

