% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEVAP.R
\encoding{UTF-8}
\name{SEVAP_ENDO}
\alias{SEVAP_ENDO}
\title{SEVAP_ENDO}
\usage{
SEVAP_ENDO(
  BP,
  TA,
  RELHUM,
  VEL,
  TC,
  TSURF,
  ELEV,
  SKINW,
  FLYHR,
  CONVSK,
  HD,
  HDFREE,
  PCTBAREVAP,
  PCTEYES,
  ZFUR,
  FURWET,
  TFA,
  CONVAR
)
}
\arguments{
\item{BP}{barometric pressure (Pa), negative means altitude is used}

\item{TA}{air temperature (째C)}

\item{RELHUM}{relative humidity (\%)}

\item{VEL}{wind speed (m/s)}

\item{TC}{core temperature (째C)}

\item{TSURF}{air temperature (째C)}

\item{ELEV}{elevation (m)}

\item{SKINW}{part of the skin surface that is wet (\%)}

\item{FLYHR}{is flight occurring this hour? (imposes forced evaporative loss)}

\item{CONVSK}{area of skin for evaporation (total skin area - hair area), m2}

\item{HD}{mass transfer coefficient}

\item{HDFREE}{free mass transfer coefficient}

\item{PCTBAREVAP}{surface area for evaporation that is skin, e.g. licking paws (\%)}

\item{PCTEYES}{surface area made up by the eye (\%) - make zero if sleeping}

\item{ZFUR}{fur depth (m)}

\item{FURWET}{part of the fur surface that is wet (\%)}

\item{TFA}{fur/air interface temperature (째C)}

\item{CONVAR}{area for convection (total area minus ventral area, as determined by PCOND), m2}
}
\description{
R wrapper for Fortran binary of SEVAP_ENDO (endotherm model)
}
