% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMULSOL.R
\encoding{UTF-8}
\name{SIMULSOL}
\alias{SIMULSOL}
\title{SIMULSOL}
\usage{
SIMULSOL(
  DIFTOL,
  IPT,
  FURVARS,
  GEOMVARS,
  ENVVARS,
  TRAITS,
  TFA,
  SKINW,
  TSKIN,
  results
)
}
\arguments{
\item{DIFTOL}{error tolerance for SIMULSOL}

\item{IPT}{geometry for SIMULSOL, 1 = cylinder, 2 = sphere, 3 = ellipsoid}

\item{FURVARS}{fur input variable vector}

\item{GEOMVARS}{shape and size input variable vector}

\item{ENVVARS}{environmental input vector}

\item{TRAITS}{other trait inputs vector}

\item{TFA}{current guess of fur/air-interface temperature (°C)}

\item{SKINW}{part of the skin surface that is wet (\%)}

\item{TSKIN}{current guess of skin temperature (°C)}
}
\description{
R wrapper for Fortran binary of SIMULSOL (endotherm model)
}
