% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOLAR.R
\encoding{UTF-8}
\name{SOLAR_ENDO}
\alias{SOLAR_ENDO}
\title{SOLAR_ENDO}
\usage{
SOLAR_ENDO(
  AREATOTL,
  ABSAND,
  ABSANV,
  ABSSB,
  ASIL,
  PCTDIF,
  QNORM,
  SHADE,
  QSOLR,
  FASKY,
  FAVEG
)
}
\arguments{
\item{AREATOTL}{surface area for solar exchange, m2}

\item{ABSAND}{solar absorptivity of dorsal fur (fractional, 0-1)}

\item{ABSANV}{solar absorptivity of ventral fur (fractional, 0-1)}

\item{ABSSB}{solar absorptivity of substrate (fractional, 0-1)}

\item{ASIL}{silhouette area to sun, m2}

\item{PCTDIF}{proportion of solar radiation that is diffuse (fractional, 0-1)}

\item{QNORM}{compute solar radiation on a surface normal to the direct rays of the sun (W/m2)}

\item{SHADE}{shade (fractional, 0-1)}

\item{QSOLR}{solar radiation, horizontal plane (W/m2)}

\item{FASKY}{configuration factor to sky (-)}

\item{FAVEG}{configuration factor to vegetation (-)}
}
\description{
R wrapper for Fortran binary of SOLAR_ENDO (endotherm model)
}
