% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZBRENT.R
\encoding{UTF-8}
\name{ZBRENT_ENDO}
\alias{ZBRENT_ENDO}
\title{ZBRENT_ENDO}
\usage{
ZBRENT_ENDO(QM1, QM2, TOL, ZBRENT.in, results)
}
\arguments{
\item{QM1}{bracketing guess 1 of metabolic heat generation, W}

\item{QM2}{bracketing guess 2 of metabolic heat generation, W}

\item{TOL}{tolerance of solution, W}

\item{ZBRENT.in}{vector of input parameters for RESPFUN}
}
\description{
R wrapper for Fortran binary of ZBRENT_ENDO (endotherm model)
}
